% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_RELIABILITY.R
\name{alpha_diagnostics}
\alias{alpha_diagnostics}
\title{Item total correlation and r drop}
\usage{
alpha_diagnostics(df)
}
\arguments{
\item{df}{dataframe with one dimension}
}
\description{
Item total correlation and r drop
}
\examples{
set.seed(12345)
df<-data.frame(matrix(.5,ncol=6,nrow=6))
correlation_martix<-as.matrix(df)
diag(correlation_martix)<-1
df<-round(generate_correlation_matrix(correlation_martix,nrows=1000),0)+5
psych::alpha(df)
alpha_diagnostics(df=df)
}
\keyword{reliability}
