% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_STRINGS.R
\name{call_to_string}
\alias{call_to_string}
\title{Model call to string}
\usage{
call_to_string(model)
}
\arguments{
\item{model}{Model object}
}
\description{
Takes a call object and convert it to string
}
\examples{
df<-generate_correlation_matrix()
model<-lm(df$X1~df$X2)
call_to_string(model)
}
\keyword{functions}
\keyword{strings}
