% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_CHECK_DATAFRAME.R
\name{cdf}
\alias{cdf}
\title{Check dataframe}
\usage{
cdf(
  df,
  name_length = (getOption("width")/3),
  digits = 2,
  nuniques = 0,
  parralel = FALSE,
  file = NULL
)
}
\arguments{
\item{df}{dataframe}

\item{name_length}{number of characters to be displayed for names}

\item{digits}{number of rounding digits}

\item{nuniques}{number of unique items to display}

\item{parralel}{if TRUE it will run using multiple cores}

\item{file}{output filename}
}
\description{
dataframe summary
}
\examples{
cdf(df=mtcars,parralel=TRUE)
cdf(df=change_data_type(mtcars,"factor"),nuniques=3)
cdf(df=data.frame(t(mtcars)),file="mtcars",nuniques=10)
cdf(df=mtcars)
cdf(df=generate_missing(mtcars))
cdf(df=infert,nuniques=10)
cdf(df=infert)
df<-data.frame(infert,
               date=seq(as.Date("2010-1-1"),
                    as.Date("2020-1-1"),
                    length.out=nrow(infert)))
cdf(df=df)
}
\keyword{check}
\keyword{dataframe}
\keyword{functions}
