% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS.R
\name{change_data_type}
\alias{change_data_type}
\title{dataframe data type transformations}
\usage{
change_data_type(df, type)
}
\arguments{
\item{df}{dataframe}

\item{type}{"character" "numeric" "factor" "factor_character" "character_factor" \cr
For "factor_character" if factors are found, are converted to characters \cr
For "character_factor" if characters are found, are converted to factors}
}
\description{
dataframe data type transformations
}
\examples{
cdf(df=change_data_type(df=mtcars,"character"))
cdf(df=change_data_type(df=mtcars,"numeric"))
cdf(df=change_data_type(df=mtcars,"factor"))
df<-change_data_type(df=mtcars,"factor")
cdf(df=change_data_type(df=df,"factor_character"))
}
\keyword{functions}
