% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS.R
\name{comparison_combinations}
\alias{comparison_combinations}
\title{Produce combinations for comparisons from dataframe names}
\usage{
comparison_combinations(df, all_orders = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{all_orders}{if TRUE the order of combination is considered 
i.e. the combination X1 X2 also appears as X2 X1 if FALSE it is assumed that X1 X2 and X2 X1 are the same and only one of them appears}
}
\description{
Produce combinations for comparisons from dataframe names
}
\examples{
comparison_combinations(generate_correlation_matrix(n=10)[,1:4])
}
\keyword{functions}
