% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_STATISTICAL.R
\name{compute_adjustment}
\alias{compute_adjustment}
\title{Compute adjustments}
\usage{
compute_adjustment(a, ntests)
}
\arguments{
\item{a}{alpha criterion}

\item{ntests}{number of tests}
}
\description{
Compute adjustments
}
\examples{
compute_adjustment(0.05,100)
}
\keyword{functions}
\keyword{statistical}
