% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPLORE_DESCRIPTIVES.R
\name{compute_aggregate}
\alias{compute_aggregate}
\title{Descriptive statistics}
\usage{
compute_aggregate(df, iv, file = NULL)
}
\arguments{
\item{df}{dataframe}

\item{iv}{index of independent variables}

\item{file}{output filename}
}
\description{
uses plyr
}
\details{
returns xlsx
}
\examples{
compute_aggregate(df=mtcars,iv=9)
compute_aggregate(df=mtcars,iv=9:10)
compute_aggregate(df=mtcars,iv=9:11)
compute_aggregate(df=mtcars,iv=9:11,file="descriptives")
}
\keyword{descriptives}
