% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPLORE_DESCRIPTIVES.R
\name{compute_crosstable}
\alias{compute_crosstable}
\title{Compute crosstables}
\usage{
compute_crosstable(df, factor_index = NULL, combinations = NULL)
}
\arguments{
\item{df}{dataframe}

\item{factor_index}{index of factors}

\item{combinations}{index of comparisons}
}
\description{
Compute crosstables
}
\examples{
combinations<-data.frame(index1=c("vs","am","gear"),index2=c("cyl","cyl","cyl"))
compute_crosstable(df=mtcars,combinations=combinations)
combinations<-data.frame(index1=c("vs","am"),index2=c("cyl","cyl"))
compute_crosstable(df=mtcars,combinations=combinations)
compute_crosstable(df=mtcars,factor_index=8:10)
}
\keyword{descriptives}
