% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_STATISTICAL.R
\name{compute_dissatenuation}
\alias{compute_dissatenuation}
\title{Compute dissatenuation}
\usage{
compute_dissatenuation(variable1, error1, variable2, error2)
}
\arguments{
\item{variable1}{vector}

\item{error1}{vector error measurement for variable1}

\item{variable2}{vector}

\item{error2}{vector error measurement for variable2}
}
\description{
Compute dissatenuation
}
\examples{
set.seed(1)
compute_dissatenuation(rnorm(10),rnorm(10),rnorm(10),rnorm(10))
}
\keyword{functions}
\keyword{statistical}
