% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_IRT_T.R
\name{compute_dummy_comparisons}
\alias{compute_dummy_comparisons}
\title{Compute number of dummy comparisons}
\usage{
compute_dummy_comparisons(items)
}
\arguments{
\item{items}{number of items per block}
}
\description{
Compute number of dummy comparisons
}
\examples{
compute_dummy_comparisons(1)
compute_dummy_comparisons(2)
compute_dummy_comparisons(3)
compute_dummy_comparisons(4)
compute_dummy_comparisons(5)
compute_dummy_comparisons(6)
}
\keyword{IRT}
\keyword{Thurstonian}
