% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPLORE_DESCRIPTIVES.R
\name{compute_frequencies}
\alias{compute_frequencies}
\title{Frequencies by levels}
\usage{
compute_frequencies(df, ordered = TRUE, file = NULL)
}
\arguments{
\item{df}{dataframe}

\item{ordered}{if TRUE it will output frequencies in descending order}

\item{file}{output filename}
}
\description{
returns frequency proportion percent
}
\details{
returns xlsx
}
\examples{
compute_frequencies(df=generate_missing(generate_factor(nrows=10,ncols=10),missing=5))
compute_frequencies(df=generate_factor())
compute_frequencies(df=generate_factor(),file="descriptives")
}
\keyword{descriptives}
