% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_IRT_U.R
\name{compute_info_1pl}
\alias{compute_info_1pl}
\title{Compute item information for 1PL model}
\usage{
compute_info_1pl(b, theta)
}
\arguments{
\item{b}{numeric difficulty parameter}

\item{theta}{numeric theta}
}
\description{
Compute item information for 1PL model
}
\examples{
compute_info_1pl(b=1,theta=-3)
compute_info_1pl(b=1,theta=-2)
compute_info_1pl(b=1,theta=-1)
compute_info_1pl(b=1,theta=0)
compute_info_1pl(b=1,theta=1)
compute_info_1pl(b=1,theta=2)
compute_info_1pl(b=1,theta=3)
ti<-compute_info_1pl(b=1,theta=seq(-6,6,by=.01)) # test information
plot(ti,x=seq(-6,6,by=.01))
}
\keyword{IRT}
\keyword{unidimensional}
