% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_IRT_U.R
\name{compute_info_2pl}
\alias{compute_info_2pl}
\title{Compute item information for 2PL model}
\usage{
compute_info_2pl(a, b, theta)
}
\arguments{
\item{a}{numeric discrimination parameter}

\item{b}{numeric difficulty parameter}

\item{theta}{numeric theta}
}
\description{
Compute item information for 2PL model
}
\examples{
compute_info_2pl(a=1.5,b=1,theta=-3)
compute_info_2pl(a=1.5,b=1,theta=-2)
compute_info_2pl(a=1.5,b=1,theta=-1)
compute_info_2pl(a=1.5,b=1,theta=0)
compute_info_2pl(a=1.5,b=1,theta=1)
compute_info_2pl(a=1.5,b=1,theta=2)
compute_info_2pl(a=1.5,b=1,theta=3)
ti<-compute_info_2pl(a=1,b=-2,theta=seq(-6,6,by=.01)) # test information
plot(ti,x=seq(-6,6,by=.01))
ti<-compute_info_2pl(a=2,b=0,theta=seq(-6,6,by=.01)) # test information
plot(ti,x=seq(-6,6,by=.01))
ti<-compute_info_2pl(a=3,b=2,theta=seq(-6,6,by=.01)) # test information
plot(ti,x=seq(-6,6,by=.01))
}
\keyword{IRT}
\keyword{unidimensional}
