% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_IRT_U.R
\name{compute_info_3pl}
\alias{compute_info_3pl}
\title{Compute item information for 3PL model}
\usage{
compute_info_3pl(a, b, g, theta)
}
\arguments{
\item{a}{numeric discrimination parameter}

\item{b}{numeric difficulty parameter}

\item{g}{numeric guessing parameter}

\item{theta}{numeric theta}
}
\description{
Compute item information for 3PL model
}
\examples{
compute_info_3pl(a=1.5,b=1,g=.2,theta=-3)
compute_info_3pl(a=1.5,b=1,g=.2,theta=-2)
compute_info_3pl(a=1.5,b=1,g=.2,theta=-1)
compute_info_3pl(a=1.5,b=1,g=.2,theta=0)
compute_info_3pl(a=1.5,b=1,g=.2,theta=1)
compute_info_3pl(a=1.5,b=1,g=.2,theta=2)
compute_info_3pl(a=1.5,b=1,g=.2,theta=3)
ti<-compute_info_3pl(a=1.5,b=1,g=.2,theta=seq(-6,6,by=.01)) # test information
plot(ti,x=seq(-6,6,by=.01))
}
\keyword{IRT}
\keyword{unidimensional}
