% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_ANOVA.R
\name{compute_kruskal_wallis_test}
\alias{compute_kruskal_wallis_test}
\title{Kruskal Wallis test}
\usage{
compute_kruskal_wallis_test(formula, df)
}
\arguments{
\item{formula}{one way formula in form of y~x. It will ignore more complex formulas}

\item{df}{dataframe}
}
\description{
Kruskal Wallis test
}
\note{
eta squared ranges between 0 and 1 \cr
epsilon squared ranges between 0 and 1 \cr
eta squared multiplied by 100 indicates the percentage of variance in the dependent variable explained by the independent variable \cr
}
\examples{
form<-formula(qsec~cyl)
kruskal.test(formula=form,data=mtcars)
rcompanion::epsilonSquared(x=mtcars$qsec,g=mtcars$cyl,group="row",ci=TRUE,conf=0.95,
                           type="perc",R=1000,digits=3)
rstatix::kruskal_effsize(mtcars,form,ci=TRUE,conf.level=0.95,ci.type="perc",nboot=100)
compute_kruskal_wallis_test(formula=form,df=mtcars)
}
\keyword{ANOVA}
