% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_STATISTICAL.R
\name{compute_kurtosis}
\alias{compute_kurtosis}
\title{Compute kurtosis}
\usage{
compute_kurtosis(vector)
}
\arguments{
\item{vector}{vector}
}
\description{
Compute kurtosis
}
\note{
b_2 = m_4 / s^4 - 3 = (g_2 + 3) (1 - 1/n)^2 - 3. Used in MINITAB and BMDP.
}
\examples{
set.seed(1)
vector<-rnorm(1000)
compute_kurtosis(vector)
e1071::kurtosis(vector)
}
\keyword{functions}
\keyword{statistical}
