% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPLORE_TIME_SERIES.R
\name{compute_moving_average}
\alias{compute_moving_average}
\title{Moving Average}
\usage{
compute_moving_average(df, w)
}
\arguments{
\item{df}{dataframe}

\item{w}{window}
}
\description{
compute moving average
}
\examples{
compute_moving_average(df=mtcars,w=5)
}
\keyword{series}
\keyword{time}
