% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_ANOVA.R
\name{compute_one_way_test}
\alias{compute_one_way_test}
\title{one way test}
\usage{
compute_one_way_test(formula, df, var.equal = TRUE)
}
\arguments{
\item{formula}{one way formula in form of y~x. It will ignore more complex formulas}

\item{df}{dataframe}

\item{var.equal}{if TRUE it assumes equal variances}
}
\description{
one way test
}
\note{
eta and omega for welch statistics are not adequatelly tested and they should not be consulted \cr
}
\examples{
form<-formula(qsec~cyl)
compute_one_way_test(formula=form,df=mtcars,var.equal=TRUE)
compute_one_way_test(formula=form,df=mtcars,var.equal=FALSE)
oneway.test(formula=form,data=mtcars,var.equal=TRUE)
oneway.test(formula=form,data=mtcars,var.equal=FALSE)
car::Anova(aov(form,data=mtcars),type=2)
model<-lm(form,data=mtcars)
lsr::etaSquared(aov(form,data=mtcars),type=3,anova=TRUE)
sjstats::anova_stats(model,digits=22)
}
\keyword{ANOVA}
