% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_ANOVA.R
\name{compute_posthoc}
\alias{compute_posthoc}
\title{Games Howell Tukey post hoc tests}
\usage{
compute_posthoc(y, x)
}
\arguments{
\item{y}{Vector continous variable}

\item{x}{Vector factor}
}
\description{
Based on http://www.psych.yorku.ca/cribbie/6130/games_howell.R
}
\examples{
compute_posthoc(y=df_blood_pressure$bp_before,x=df_blood_pressure$agegrp)
compute_posthoc(y=df_blood_pressure$bp_after,x=df_blood_pressure$agegrp)
}
\keyword{ANOVA}
