% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_CORRELATION.R
\name{compute_power_r}
\alias{compute_power_r}
\title{Compute r power curve}
\usage{
compute_power_r(
  n = 100,
  r = NULL,
  sig.level = 0.05,
  alternative = c("two.sided", "less", "greater"),
  title = "",
  base_size = 10
)
}
\arguments{
\item{n}{number of observations}

\item{r}{correlation coefficient}

\item{sig.level}{alpha (type I error probability)}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less"}

\item{title}{plot title}

\item{base_size}{base font size}
}
\description{
Compute r power curve
}
\examples{
compute_power_r(n=100,r=.5,sig.level=.05,alternative=c("two.sided"))
}
\keyword{correlation}
