% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_CORRELATION.R
\name{compute_power_r_matrix}
\alias{compute_power_r_matrix}
\title{Compute correlation matrix}
\usage{
compute_power_r_matrix(m, ...)
}
\arguments{
\item{m}{correlation matrix}

\item{...}{arguments passed to compute_power_r}
}
\description{
Compute correlation matrix
}
\examples{
compute_power_r_matrix(m=stats::cor(mtcars,use="pairwise.complete.obs"),n=100)
}
\keyword{correlation}
