% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_EFA.R
\name{compute_residual_stats}
\alias{compute_residual_stats}
\title{Residuals for matrices}
\usage{
compute_residual_stats(model, data = NULL)
}
\arguments{
\item{model}{psych EFA model. It has to be a correlation or covariance matrix if data is not NULL}

\item{data}{correlation or covariance matrix}
}
\description{
Root Mean Squared Residual Number of absolute residuals > 0.05  Proportion of absolute residuals > 0.05. It can either accept a psych EFA model or it can compare two correlation or covariance matrices
}
\examples{
model<-psych::fa(mtcars,nfactors=2,rotate="oblimin",fm="pa",oblique.scores=TRUE)
compute_residual_stats(model)
}
\keyword{EFA}
