% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_IRT_U.R
\name{compute_se_theta}
\alias{compute_se_theta}
\title{Compute the SE of theta}
\usage{
compute_se_theta(info)
}
\arguments{
\item{info}{numeric information}
}
\description{
Compute the SE of theta
}
\examples{
compute_se_theta(1)
ti<-compute_info_2pl(a=10,b=0,theta=seq(-3,3,by=.01)) # test information
plot(compute_se_theta(ti),x=seq(-3,3,by=.01))
}
\keyword{IRT}
\keyword{unidimensional}
