% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_STATISTICAL.R
\name{compute_skewness}
\alias{compute_skewness}
\title{Compute skewness}
\usage{
compute_skewness(vector)
}
\arguments{
\item{vector}{vector}
}
\description{
Compute skewness
}
\note{
b_1 = m_3 / s^3 = g_1 ((n-1)/n)^(3/2). Used in MINITAB and BMDP.
}
\examples{
set.seed(1)
vector<-rnorm(1000)
compute_skewness(vector)
e1071::skewness(vector)
}
\keyword{functions}
\keyword{statistical}
