% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_STATISTICAL.R
\name{compute_standard}
\alias{compute_standard}
\title{compute standard scores}
\usage{
compute_standard(vector, mean = 0, sd = 1, type = "z", input = "non_standard")
}
\arguments{
\item{vector}{vector}

\item{mean}{numeric applicable to "uz"}

\item{sd}{numeric applicable to "uz"}

\item{type}{"z" "uz" "sten" "t" "stanine" "center" "center_reversed" "percent" "scale_zero_one" "normal_density" "cumulative_density" "all"}

\item{input}{"standard" "non_standard" standard inputs are z scores and non standard are raw scores}
}
\description{
compute standard scores
}
\examples{
vector<-c(rnorm(10),NA,rnorm(10))
compute_standard(vector,type="z")
compute_standard(vector,mean=0,sd=1,type="uz")
compute_standard(vector,type="sten")
compute_standard(vector,type="t")
compute_standard(vector,type="stanine")
compute_standard(vector,type="center")
compute_standard(vector,type="center_reversed")
compute_standard(vector,type="percent")
compute_standard(vector,type="scale_zero_one")
ndf<-compute_standard(seq(-6,6,.01),mean=0,sd=1,type="normal_density")
plot(ndf)
cdf<-compute_standard(ndf,mean=0,sd=1,type="cumulative_density")
plot(cdf)
compute_standard(vector,type="all")
compute_standard(seq(-6,6,.1),type="all",input="standard")
}
\keyword{functions}
\keyword{statistical}
