% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_STATISTICAL.R
\name{compute_standard_error}
\alias{compute_standard_error}
\title{Compute standard error}
\usage{
compute_standard_error(vector)
}
\arguments{
\item{vector}{vector}
}
\description{
Compute standard error
}
\examples{
set.seed(1)
vector<-rnorm(1000)
compute_standard_error(vector)
}
\keyword{functions}
\keyword{statistical}
