% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_IRT_U.R
\name{compute_unidimensional_ability}
\alias{compute_unidimensional_ability}
\title{Compute theta for unidimensional models}
\usage{
compute_unidimensional_ability(
  a,
  b,
  g = NULL,
  d = 1.702,
  u,
  lim_theta = c(-6, 6)
)
}
\arguments{
\item{a}{numeric vector discrimination parameters}

\item{b}{numeric vector difficulty parameters}

\item{g}{numeric vector guessing parameters}

\item{d}{numeric scaling constant usually it is a value that approximating 1.749}

\item{u}{numeric vector responses}

\item{lim_theta}{vector minimum and maximum value of theta}
}
\description{
Compute theta for unidimensional models
}
\examples{
a<-c(0.39,0.45,0.52,0.3,0.35,0.43,0.42,0.44,0.34,0.42)
b<-c(-1.96,-1.9,-1.38,-0.58,0.48,-0.81,-0.35,1.59,1.33,2.93)
u<-c(1,1,1,1,0,0,1,0,1,0)
# SHOULD RETURN 0.48402574251176
compute_unidimensional_ability(a=a,b=b,u=u,d=1.7,g=NULL)
a<-c(1.27,0.9,0.94,0.95,0.55,0.6,0.44,0.4)
b<-c(-0.54,0.18,0.21,1.26,1.73,-0.87,1.72,2.67)
u<-c(1,1,1,1,0,0,0,0)
# SHOULD RETURN 1.04621621510192
compute_unidimensional_ability(a=a,b=b,u=u,d=1.7,g=NULL)
a<-c(0.41,0.32,0.33,1.2,0.63,0.62,0.7,0.61,0.38,0.53,0.6,1.16)
b<-c(-1.4,-1.3,-1.17,0.2,0.71,0.86,-0.12,0.12,2.06,1.38,1.18,-0.33)
u<-c(1,0,1,1,0,0,0,1,1,0,1,0)
# SHOULD RETURN 0.0860506282671103
compute_unidimensional_ability(a=a,b=b,u=u,d=1.7,g=NULL)
}
\keyword{IRT}
\keyword{unidimensional}
