% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_IRT_U.R
\name{compute_unidimensional_theta}
\alias{compute_unidimensional_theta}
\title{Compute theta for unidimensional models}
\usage{
compute_unidimensional_theta(a, b = 0, g = 0, i = 1, d = 1.702, theta = 0)
}
\arguments{
\item{a}{numeric discrimination parameter}

\item{b}{numeric difficulty parameter}

\item{g}{numeric guessing parameter}

\item{i}{numeric innatentiveness parameter}

\item{d}{numeric scaling constant usually a value 1.749 or 1.702}

\item{theta}{numeric or vector theta}
}
\description{
Compute theta for unidimensional models
}
\note{
when scaling constant=1 it has no effect in equation\cr
      when innatentiveness=1 and guessing=0 function computes a 2PL score\cr
      when innatentiveness=1 and guessing!=0 function computes a 3PL score\cr
      when innatentiveness!=1 and guessing!=0 function computes a 4PL score\cr
}
\examples{
compute_unidimensional_theta(a=10,b=0)
x<-seq(-3,3,by=.01)
plot(compute_unidimensional_theta(a=5,b=0,theta=x),x=x)
plot(compute_unidimensional_theta(a=5,b=-1,theta=x),x=x)
plot(compute_unidimensional_theta(a=5,b=1,theta=x),x=x)
plot(compute_unidimensional_theta(a=.1,b=0,theta=x),x=x)
plot(compute_unidimensional_theta(a=1,b=0,theta=x),x=x)
plot(compute_unidimensional_theta(a=10,b=0,theta=x),x=x)
plot(compute_unidimensional_theta(a=10,b=0,g=0,theta=x),x=x)
plot(compute_unidimensional_theta(a=10,b=0,g=.1,theta=x),x=x)
plot(compute_unidimensional_theta(a=10,b=0,g=.5,theta=x),x=x)
plot(compute_unidimensional_theta(a=10,b=0,g=0,i=1,theta=x),x=x)
plot(compute_unidimensional_theta(a=10,b=0,g=0,i=.9,theta=x),x=x)
plot(compute_unidimensional_theta(a=10,b=0,g=0,i=.6,theta=x),x=x)
}
\keyword{IRT}
\keyword{unidimensional}
