% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_LOGISTIC_REGRESSION.R
\name{compute_y_logistic}
\alias{compute_y_logistic}
\title{Compute y for logistic function}
\usage{
compute_y_logistic(intercept, coefficient, x)
}
\arguments{
\item{intercept}{Numeric}

\item{coefficient}{Numeric}

\item{x}{Numeric}
}
\description{
This function requires x range to produce a vector with y values
}
\examples{
x<--10:10
compute_y_logistic(0,1,x)
compute_y_logistic(0,1,1)
plot(x,compute_y_logistic(0,1,x),type="l");grid();abline(b=0,a=.5)
}
\keyword{logistic}
\keyword{regression}
