% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_UNIX_TIME.R
\name{convert_excel_unix_timestamp}
\alias{convert_excel_unix_timestamp}
\title{Convert UNIX EXCEL timestamp}
\usage{
convert_excel_unix_timestamp(timestamp)
}
\arguments{
\item{timestamp}{unix or excel timestamp}
}
\description{
Convert UNIX EXCEL timestamp
}
\examples{
convert_excel_unix_timestamp(1)
}
\keyword{timestamp}
