% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_MATHEMATICAL.R
\name{deg2rad}
\alias{deg2rad}
\title{Convert degrees to radians}
\usage{
deg2rad(degrees)
}
\arguments{
\item{degrees}{degrees}
}
\description{
Convert degrees to radians
}
\examples{
deg2rad(180)
}
\keyword{correlation}
