% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{df_admission}
\alias{df_admission}
\title{Admission Data}
\format{
A data frame with 8 rows and 4 variables:
\describe{
  \item{admit}{Binary variable indicating admission (0 = No, 1 = Yes)}
  \item{gre}{GRE (Graduate Record Examination) score}
  \item{gpa}{Grade Point Average}
  \item{rank}{Ranking of the undergraduate institution (1 = highest, 4 = lowest)}
}
}
\source{
researchpy repo
}
\usage{
df_admission
}
\description{
This data set contains information about graduate admission, including GRE scores, GPA, and the ranking of the undergraduate institution.
}
\keyword{datasets}
