% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{df_blood_pressure}
\alias{df_blood_pressure}
\title{Blood Pressure Data}
\format{
A data frame with 30 rows and 5 variables:
\describe{
  \item{patient}{Unique identifier for each patient}
  \item{sex}{Sex of the patient (e.g., Male, Female)}
  \item{agegrp}{Age group of the patient (e.g., 30-45, 46-59)}
  \item{bp_before}{Blood pressure reading before the intervention}
  \item{bp_after}{Blood pressure reading after the intervention}
}
}
\source{
researchpy repo
}
\usage{
df_blood_pressure
}
\description{
This data set contains blood pressure readings for patients before and after a certain treatment or intervention.
}
\keyword{datasets}
