% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{df_personality}
\alias{df_personality}
\title{Personality Dataset}
\format{
A data frame with 22 rows and 44 variables:
\describe{
  \item{pers01}{Personality item 1}
  \item{pers02}{Personality item 2}
  \item{pers03}{Personality item 3}
  \item{pers04}{Personality item 4}
  \item{pers05}{Personality item 5}
  \item{pers06}{Personality item 6}
  \item{pers07}{Personality item 7}
  \item{pers08}{Personality item 8}
  \item{pers09}{Personality item 9}
  \item{pers10}{Personality item 10}
  \item{pers11}{Personality item 11}
  \item{pers12}{Personality item 12}
  \item{pers13}{Personality item 13}
  \item{pers14}{Personality item 14}
  \item{pers15}{Personality item 15}
  \item{pers16}{Personality item 16}
  \item{pers17}{Personality item 17}
  \item{pers18}{Personality item 18}
  \item{pers19}{Personality item 19}
  \item{pers20}{Personality item 20}
  \item{pers21}{Personality item 21}
  \item{pers22}{Personality item 22}
  \item{pers23}{Personality item 23}
  \item{pers24}{Personality item 24}
  \item{pers25}{Personality item 25}
  \item{pers26}{Personality item 26}
  \item{pers27}{Personality item 27}
  \item{pers28}{Personality item 28}
  \item{pers29}{Personality item 29}
  \item{pers30}{Personality item 30}
  \item{pers31}{Personality item 31}
  \item{pers32}{Personality item 32}
  \item{pers33}{Personality item 33}
  \item{pers34}{Personality item 34}
  \item{pers35}{Personality item 35}
  \item{pers36}{Personality item 36}
  \item{pers37}{Personality item 37}
  \item{pers38}{Personality item 38}
  \item{pers39}{Personality item 39}
  \item{pers40}{Personality item 40}
  \item{pers41}{Personality item 41}
  \item{pers42}{Personality item 42}
  \item{pers43}{Personality item 43}
  \item{pers44}{Personality item 44}
}
}
\usage{
df_personality
}
\description{
A dataset containing personality test results.
}
\examples{
data(df_personality)
head(df_personality)
}
\keyword{datasets}
