% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{df_responses}
\alias{df_responses}
\title{Responses Data}
\format{
A data frame with [number of rows] rows and [number of variables] variables.
}
\source{
Downloaded from Kaggle.com by the user Miroslav Sabo. \url{https://www.kaggle.com/miroslavsabo/young-people-survey}
}
\usage{
df_responses
}
\description{
This data set contains responses data.
}
\keyword{datasets}
