% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{df_sexual_comp}
\alias{df_sexual_comp}
\title{Sexual Compatibility Data}
\format{
A data frame with 22 rows and 13 variables:
\describe{
  \item{Q1}{Response to question 1}
  \item{Q2}{Response to question 2}
  \item{Q3}{Response to question 3}
  \item{Q4}{Response to question 4}
  \item{Q5}{Response to question 5}
  \item{Q6}{Response to question 6}
  \item{Q7}{Response to question 7}
  \item{Q8}{Response to question 8}
  \item{Q9}{Response to question 9}
  \item{Q10}{Response to question 10}
  \item{score}{Total score}
  \item{gender}{Gender of the respondent (1 = Male, 2 = Female)}
  \item{age}{Age of the respondent}
}
}
\source{
researchpy repo
}
\usage{
df_sexual_comp
}
\description{
This data set contains responses to questions about sexual compatibility, 
including scores, gender, and age.
}
\keyword{datasets}
