% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{df_titanic}
\alias{df_titanic}
\title{Titanic Dataset}
\format{
A data frame with the following variables:
\describe{
  \item{PassengerId}{Unique identifier for each passenger}
  \item{survived}{Survival status (0 = No, 1 = Yes)}
  \item{pclass}{Passenger class (1 = 1st, 2 = 2nd, 3 = 3rd)}
  \item{name}{Name of the passenger}
  \item{sex}{Gender of the passenger}
  \item{age}{Age of the passenger}
  \item{sibsp}{Number of siblings/spouses aboard the Titanic}
  \item{parch}{Number of parents/children aboard the Titanic}
  \item{ticket}{Ticket number}
  \item{fare}{Passenger fare}
  \item{cabin}{Cabin number}
  \item{embarked}{Port of embarkation (C = Cherbourg; Q = Queenstown; S = Southampton)}
  \item{boat}{Lifeboat number}
  \item{body}{Body number}
  \item{home.dest}{Home destination}
}
}
\usage{
df_titanic
}
\description{
A dataset containing information about passengers on the Titanic.
}
\examples{
data(df_titanic)
head(df_titanic)
}
\keyword{datasets}
