% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS.R
\name{dotnames}
\alias{dotnames}
\title{Get the names of objects in the arguments}
\usage{
dotnames(...)
}
\arguments{
\item{...}{objects}
}
\description{
Get the names of objects in the arguments
}
\author{
Ananda Mahto
}
\keyword{functions}
