% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_RECODE.R
\name{drop_levels}
\alias{drop_levels}
\title{Drops unused factor levels}
\usage{
drop_levels(df, factor_index = NULL, minimum_frequency = 5)
}
\arguments{
\item{df}{dataframe}

\item{factor_index}{numeric index of factors. If NULL the function uses is.factor() to discriminate factors}

\item{minimum_frequency}{the minimum frequency each factor will have, levels with frequency bellow or equal to the defined frequency will be renamed "Other"}
}
\description{
Drops unused factor levels
}
\examples{
factor1<-factor(c(rep("A",10),rep("B",10)),levels=c("A","B","C","D"))
factor2<-factor(c(rep("A",10),rep("B",10)),levels=c("A","B","C","D"))
numeric1<-c(1:20)
df<-data.frame(numeric1,factor1,factor2)
df$factor1
drop_levels(df=df,minimum_frequency=9)
drop_levels(df=df,minimum_frequency=10)
}
\keyword{functions}
