% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_RECODE.R
\name{dummy_arrange}
\alias{dummy_arrange}
\title{Takes a vector with multiple responses and dummy arranges it in a dataframe}
\usage{
dummy_arrange(vector)
}
\arguments{
\item{vector}{Vector}
}
\description{
Takes a vector with multiple responses and dummy arranges it in a dataframe
}
\examples{
vector1<-gsub(" ","",
             generate_multiple_responce_vector(responces=c("Agree","Hi","All"),
             responded=1:3,length=10),fixed=TRUE)
vector2<-gsub(" ","",
             generate_multiple_responce_vector(responces=1:4,responded=1:4,length=10),
             fixed=TRUE)
vector3<-sample(1:4,10,replace=TRUE)
vector4<-sample(LETTERS[1:3],10,replace=TRUE)
dummy_arrange(vector1)
dummy_arrange(vector2)
dummy_arrange(vector3)
dummy_arrange(vector4)
}
\keyword{functions}
