% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_PLOT.R
\name{duplicate_y_axis}
\alias{duplicate_y_axis}
\title{Duplicate y axis}
\usage{
duplicate_y_axis(p1, p2)
}
\arguments{
\item{p1}{Plot 1}

\item{p2}{Plot 2}
}
\description{
Duplicate y axis
}
\examples{
p1<-ggplot(ChickWeight,aes(x=Time,y=weight,colour=Diet,group=Chick))+
           geom_line()+
           ggtitle("Growth curve for individual chicks")
duplicate_y_axis(p1,p1)
}
\keyword{functions}
\keyword{plot}
