% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_ENVIRONMENT.R
\name{environment_options}
\alias{environment_options}
\title{Load environment options}
\usage{
environment_options()
}
\description{
Load environment options
}
\examples{
environment_options()
}
\keyword{functions}
