% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_EXCEL.R
\name{excel_confusion_matrix}
\alias{excel_confusion_matrix}
\title{Write matrix or dataframe to excel sheet}
\usage{
excel_confusion_matrix(
  df,
  workbook,
  title = "Rows: Expected Collumns: Observed"
)
}
\arguments{
\item{df}{dataframe or matrix}

\item{workbook}{workbook}

\item{title}{comment}
}
\description{
Usefull for correlation matrices since it uses conditional formatting for matrices
}
\examples{
filename<-"excel_confusion_matrix.xlsx"
if (file.exists(filename)) file.remove(filename)
observed<-factor(round(rnorm(10000,m=10,sd=1)))
predicted<-factor(round(rnorm(10000,m=10,sd=1)))
confusion(observed,predicted)
cm<-confusion_matrix_percent(observed,predicted)
wb<-openxlsx::createWorkbook()
excel_confusion_matrix(cm,wb)
openxlsx::saveWorkbook(wb,invisible(paste(filename)),TRUE)
}
\keyword{functions}
