% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_RELIABILITY.R
\name{extract_components}
\alias{extract_components}
\title{Extract variance components from model}
\usage{
extract_components(model, title = "")
}
\arguments{
\item{model}{model containing variance components}

\item{title}{plot title}
}
\description{
Extract variance components from model
}
\examples{
design<-expand.grid(time=1:3,item=1:3,person=1:10)
design<-change_data_type(design,type="factor")
design$response<-rowSums(change_data_type(design[,1:2],type="numeric"))+rnorm(90,0,0.1)
model<-mixlm::lm(response~r(time)*r(person)+r(item)*r(person),data=design)
extract_components(model)
}
\keyword{reliability}
