% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_RECODE.R
\name{flatten_list}
\alias{flatten_list}
\title{Flatten two dimensional list}
\usage{
flatten_list(mydata)
}
\arguments{
\item{mydata}{list with two dimensions}
}
\description{
Flatten two dimensional list
}
\keyword{functions}
