% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_GENERATEDATA.R
\name{generate_correlation_matrix}
\alias{generate_correlation_matrix}
\title{Generate dataframe which outputs a predetermined correlation matrix}
\usage{
generate_correlation_matrix(correlation_martix, nrows = 10)
}
\arguments{
\item{correlation_martix}{correlation matrix of resulting dataframe}

\item{nrows}{number of rows to generate}
}
\description{
Generate dataframe which outputs a predetermined correlation matrix
}
\examples{
df<-data.frame(matrix(.999,ncol=2,nrow=2))
correlation_martix<-as.matrix(df)
diag(correlation_martix)<-1
df<-generate_correlation_matrix(correlation_martix,nrows=100)
stats::cor(df)
}
\keyword{data}
\keyword{functions}
\keyword{generate}
