% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_GENERATEDATA.R
\name{generate_data}
\alias{generate_data}
\title{Generate dataframe with random numbers}
\usage{
generate_data(
  nrows = 10,
  ncols = 5,
  mean = 0,
  sd = 1,
  min = 1,
  max = 5,
  type = "normal"
)
}
\arguments{
\item{nrows}{number of rows to generate}

\item{ncols}{number of collumns to generate}

\item{mean}{mean of generated vectors}

\item{sd}{standard deviation of generated vectors}

\item{min}{minimum value in generated vector}

\item{max}{maximum value in generated vector}

\item{type}{character "normal" "uniform"}
}
\description{
Generate dataframe with random numbers
}
\examples{
generate_data(nrows=10,ncols=5,mean=0,sd=1,type="normal")
generate_data(nrows=10,ncols=5,min=1,max=5,type="uniform")
}
\keyword{data}
\keyword{functions}
\keyword{generate}
