% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_GENERATEDATA.R
\name{generate_factor}
\alias{generate_factor}
\title{Generate dataframe of factors}
\usage{
generate_factor(vector = LETTERS[1:5], nrows = 2, ncols = 10, type = "random")
}
\arguments{
\item{vector}{factor pool}

\item{nrows}{number of rows to generate}

\item{ncols}{number of collumns to generate}

\item{type}{"balanced" or "random" "balanced" generates balanced factor vectrors, "random" generates random factor vectors}
}
\description{
Generate dataframe of factors
}
\examples{
generate_factor(vector=LETTERS[1:5],ncols=5,nrows=10,type="random")
generate_factor(vector=LETTERS[1:5],ncols=5,nrows=10,type="balanced")
generate_factor(vector=LETTERS[1:5],ncols=1,nrows=10,type="balanced")
generate_factor(vector=LETTERS[1:5],ncols=1,nrows=10,type="random")
}
\keyword{data}
\keyword{functions}
\keyword{generate}
