% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_GENERATEDATA.R
\name{generate_missing}
\alias{generate_missing}
\title{Generate missing data}
\usage{
generate_missing(df, missing = 5)
}
\arguments{
\item{df}{vector or dataframe}

\item{missing}{number of missing data per vector}
}
\description{
Generate missing data
}
\examples{
generate_missing(rnorm(10),missing=5)
generate_missing(generate_data(nrow=10,ncol=2),missing=5)
}
\keyword{data}
\keyword{functions}
\keyword{generate}
