% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_GENERATEDATA.R
\name{generate_multiple_responce_vector}
\alias{generate_multiple_responce_vector}
\title{Generate multiple responce vector}
\usage{
generate_multiple_responce_vector(
  responces = 1:4,
  responded = 1:4,
  length = 10
)
}
\arguments{
\item{responces}{unique categories allowed}

\item{responded}{number of categories observed in iteration}

\item{length}{length of returned vector}
}
\description{
Generate multiple responce vector
}
\examples{
generate_multiple_responce_vector(responces=1:4,responded=1:4,length=10)
}
\keyword{data}
\keyword{functions}
\keyword{generate}
