% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_GENERATEDATA.R
\name{generate_string}
\alias{generate_string}
\title{Generate random strings}
\usage{
generate_string(
  vector = c(LETTERS, letters, 0:9),
  vector_length = 1,
  nchar = 5
)
}
\arguments{
\item{vector}{character pool}

\item{vector_length}{number of strings to generate}

\item{nchar}{Length of generated strings}
}
\description{
Generate random strings
}
\examples{
generate_string(nchar=10)
generate_string(nchar=10,vector_length=10)
}
\keyword{data}
\keyword{functions}
\keyword{generate}
