% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_RECODE.R
\name{get_coding}
\alias{get_coding}
\title{Takes question response options from a json object and returns a list of dataframes mapping response option number to response option text}
\usage{
get_coding(coding_list)
}
\arguments{
\item{coding_list}{Json object containing question options}
}
\description{
Takes question response options from a json object and returns a list of dataframes mapping response option number to response option text
}
\examples{
file<-"/opt/repo/rworking/projects/pcentral/data/aristotle-accountests-6-17.json"
coding_list<-rjson::fromJSON(file=file)$instr_descr_options$biodata
get_coding(coding_list)
}
\keyword{functions}
