% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_IRT_T.R
\name{get_mplus_thu_3t}
\alias{get_mplus_thu_3t}
\title{Simulate prior distribution}
\usage{
get_mplus_thu_3t(model)
}
\arguments{
\item{model}{mplus thurstonian cfa model with 3 traits}
}
\description{
Simulate prior distribution
}
\keyword{IRT}
\keyword{Thurstonian}
