% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_KEYS.R
\name{get_questionaire_reference}
\alias{get_questionaire_reference}
\title{Questionnaire reference}
\usage{
get_questionaire_reference(mydata, questionnaire, tag_remove = "")
}
\arguments{
\item{mydata}{json output}

\item{questionnaire}{name of questionnaire to extract reference from}

\item{tag_remove}{character}
}
\description{
Questionnaire reference
}
\examples{
stud<-rjson::fromJSON(file="/opt/repo/rworking/projects/jacob/stud.json")
iboSTUDPULS2<-json2dataframe(stud$`items-by-order`$STUDPULS2)
get_questionaire_reference(stud,"STUDPULS2","DTL2-ST")
}
\keyword{dimensions}
\keyword{functions}
\keyword{keys}
