% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_KEYS.R
\name{get_reversals}
\alias{get_reversals}
\title{Questionnaire reference}
\usage{
get_reversals(df)
}
\arguments{
\item{df}{dataframe from get_questionaire_reference}
}
\description{
Questionnaire reference
}
\examples{
stud<-rjson::fromJSON(file="/opt/repo/rworking/projects/jacob/stud.json")
res<-get_questionaire_reference(stud,"STUDPULS2","DTL2-ST")
ibo<-res$items_by_order[!res$items_by_order$instrdtl_value \%in\% "-9999.00",]
ibo$instrdtl_value<-as.numeric(ibo$instrdtl_value)
get_reversals(ibo)
}
\keyword{dimensions}
\keyword{functions}
\keyword{keys}
