% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_PLOT.R
\name{hinvert_title_grob}
\alias{hinvert_title_grob}
\title{Invert title for duplicating y axis}
\usage{
hinvert_title_grob(grob)
}
\arguments{
\item{grob}{grob object}
}
\description{
Invert title for duplicating y axis
}
\keyword{functions}
\keyword{plot}
