% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_IRT_T.R
\name{increase_index}
\alias{increase_index}
\title{index dataframe picks}
\usage{
increase_index(blocks, items)
}
\arguments{
\item{blocks}{number of blocks}

\item{items}{number of items per block}
}
\description{
index dataframe picks
}
\examples{
increase_index(3,3)
}
\keyword{IRT}
\keyword{Thurstonian}
