% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_ENVIRONMENT.R
\name{install_all_packages}
\alias{install_all_packages}
\title{Install all packages available in CRAN}
\usage{
install_all_packages()
}
\description{
Install all packages available in CRAN
}
\details{
Install all packages available in CRAN. Already installed packages are not downloaded or installed
}
\keyword{functions}
