% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_ENVIRONMENT.R
\name{install_load}
\alias{install_load}
\title{Install and load multiple packages}
\usage{
install_load(package)
}
\arguments{
\item{package}{Vector Package names}
}
\description{
Install and load multiple packages. If packages exist,they are loaded,if packages don't exist,they are downloaded installed and loaded
}
\examples{
install_load("car")
install_load(c("car","ggplot2"))
}
\author{
Steven Worthington
}
\keyword{functions}
