% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_RECODE.R
\name{json2dataframe}
\alias{json2dataframe}
\title{Convert a json object to dataframe}
\usage{
json2dataframe(mydata, fast = FALSE, row_to_collumn = FALSE)
}
\arguments{
\item{mydata}{json object}

\item{fast}{if TRUE uses a faster algorythm}

\item{row_to_collumn}{if TRUE it will return row in first collumn}
}
\description{
Convert a json object to dataframe
}
\examples{
file<-"/opt/repo/rworking/projects/pcentral/data/aristotle-accountests-6-17.json"
pcentral<-rjson::fromJSON(file=file)
biodata<-json2dataframe(mydata=pcentral$itemresponses$biodata)
biodata<-json2dataframe(mydata=pcentral$itemresponses$biodata,fast=TRUE)
}
\keyword{functions}
