% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_TRAIN_TEST.R
\name{k_sample}
\alias{k_sample}
\title{train test sampling}
\usage{
k_sample(df, model_formula, k = 1)
}
\arguments{
\item{df}{dataframe}

\item{model_formula}{model formula}

\item{k}{k-folds}
}
\description{
splits a dataframe in train and test dataframes for model evaluation. Prepared data include data objects for xgboost
}
\examples{
infert_formula<-as.formula(factor(case)~age+parity+education+spontaneous+induced)
result<-k_sample(df=infert,k=10,model_formula=infert_formula)
model_formula<-as.formula(mpg~cyl+disp+hp+drat+wt+qsec+vs+am+gear+carb)
result<-k_sample(df=mtcars,k=10,model_formula=model_formula)
}
\keyword{functions}
