% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_RELIABILITY.R
\name{mean_sd_alpha}
\alias{mean_sd_alpha}
\title{Mean and SD}
\usage{
mean_sd_alpha(df, divisor = NULL)
}
\arguments{
\item{df}{dataframe with one dimension}

\item{divisor}{number to use for dividing the rowsums}
}
\description{
Mean and SD
}
\examples{
set.seed(12345)
df<-data.frame(matrix(.5,ncol=6,nrow=6))
correlation_martix<-as.matrix(df)
diag(correlation_martix)<-1
df<-round(generate_correlation_matrix(correlation_martix,nrows=1000),0)+5
mean_sd_alpha(df)
mean_sd_alpha(df,divisor=100)
}
\keyword{reliability}
