% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_STRINGS.R
\name{mgsub}
\alias{mgsub}
\title{Sub for multiple patterns}
\usage{
mgsub(mydata, pattern, replacement, ...)
}
\arguments{
\item{mydata}{Character}

\item{pattern}{Character to search for}

\item{replacement}{Replacement character}

\item{...}{arguments passed to gsub}
}
\description{
Sub for multiple patterns
}
\examples{
mgsub(mydata="#$\%^&*_+",pattern=c("\%","*"),"REPLACE",fixed=TRUE)
}
\keyword{functions}
\keyword{strings}
