% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_EFA.R
\name{model_loadings}
\alias{model_loadings}
\title{Pattern and structure matrix}
\usage{
model_loadings(model, cut = NULL, matrix_type = "pattern", sort = TRUE, ...)
}
\arguments{
\item{model}{psych EFA model}

\item{cut}{cut point for loadings}

\item{matrix_type}{"pattern" "structure" "all"}

\item{sort}{if TRUE it will sort loadings}

\item{...}{arguments passed to psych::fa.sort}
}
\description{
Pattern and structure matrix
}
\note{
Check to see if you have multicolinearity values above .8 in the matrix are problematic \cr
Structure matrix represents Loadings after rotation \cr
Pattern matrix represents Loadings before rotation \cr
}
\examples{
model<-psych::fa(mtcars,nfactors=2,rotate="oblimin",fm="pa",oblique.scores=TRUE)
model_loadings(model=model,cut=NULL,matrix_type="pattern")
model_loadings(model=model,cut=0.4,matrix_type="structure")
model_loadings(model=model,cut=0.4,matrix_type="all",sort=FALSE)
}
\keyword{EFA}
