% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_MATRIX.R
\name{off_diagonal_index}
\alias{off_diagonal_index}
\title{index of off diagonal}
\usage{
off_diagonal_index(length)
}
\arguments{
\item{length}{length of diagonal}
}
\description{
index of off diagonal
}
\examples{
off_diagonal_index(length=6)
}
\keyword{functions}
\keyword{matrix}
