% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPLORE_ASSUMPTIONS.R
\name{outlier_summary}
\alias{outlier_summary}
\title{Percent of outliers in vector}
\usage{
outlier_summary(vector)
}
\arguments{
\item{vector}{numeric vector}
}
\description{
Percent of outliers in vector
}
\details{
returns dataframe
}
\examples{
vector<-generate_missing(rnorm(1000))
df<-generate_missing(mtcars[,1:2])
outlier_summary(vector)
data.frame(sapply(mtcars,outlier_summary))
}
\keyword{assumptions}
