% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_LOGISTIC_REGRESSION.R
\name{output_compare_model_logistic}
\alias{output_compare_model_logistic}
\title{Compare logistic regression models models}
\usage{
output_compare_model_logistic(model1, model2)
}
\arguments{
\item{model1}{object glm model}

\item{model2}{object glm model}
}
\description{
Compare logistic regression models models
}
\examples{
modelcategoricalpredictor<-glm(case~education,data=infert,family=binomial)
modelcontinuouspredictor<-glm(case~age,data=infert,family=binomial)
modeltwopredictors<-glm(case~education*age,data=infert,family=binomial)
modelmultiple<-glm(case~education*age*parity,data=infert,family=binomial)
anova(modelcategoricalpredictor,modelcontinuouspredictor)
output_compare_model_logistic(model1=modelcategoricalpredictor,
                              model2=modeltwopredictors)
output_compare_model_logistic(model1=modelcontinuouspredictor,
                              model2=modeltwopredictors)
output_compare_model_logistic(model1=modelcontinuouspredictor,
                              model2=modelcategoricalpredictor)
}
\keyword{logistic}
\keyword{regression}
