% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_STRINGS.R
\name{output_separator}
\alias{output_separator}
\title{Output separator}
\usage{
output_separator(
  string,
  output = NULL,
  instruction = NULL,
  length = getOption("width")/2
)
}
\arguments{
\item{string}{Title of output}

\item{output}{object to print}

\item{instruction}{Character provided instructions regarding the output}

\item{length}{Numeric Length of separator measured in number of characters}
}
\description{
Heading, main output, and instructions for output for the console environment
}
\examples{
output_separator(string="TEST",output="TEST",instruction="TEST",length=100)
output_separator(string="TEST",instruction="TEST",length=100)
output_separator(string="TEST",output="TEST",length=100)
output_separator(string="TEST")
}
\keyword{functions}
\keyword{strings}
