% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS.R
\name{padNA}
\alias{padNA}
\title{pad NA's to collumns in dataframe}
\usage{
padNA(df, rowsneeded, first = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{rowsneeded}{Numeric number of rows needed}

\item{first}{Boolean}
}
\description{
pad NA's to collumns in dataframe
}
\author{
Ananda Mahto
}
\keyword{functions}
