% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPLORE_TIME_SERIES.R
\name{plot_acf}
\alias{plot_acf}
\title{Plot autocorrelation function of correlation covariance and partial correlation}
\usage{
plot_acf(df, lag.max = length(df), base_size = 10, title = "")
}
\arguments{
\item{df}{ts object}

\item{lag.max}{maximum lags to include}

\item{base_size}{base font size}

\item{title}{plot title}
}
\description{
uses ggplot
}
\details{
returns plot
}
\examples{
ts_data<-ts(UKDriverDeaths,start=1969,end=1984,frequency=12)
plot_acf(df=ts_data,base_size=20)
}
\keyword{series}
\keyword{time}
