% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPLORE_ASSUMPTIONS.R
\name{plot_boxplot}
\alias{plot_boxplot}
\title{Boxplot}
\usage{
plot_boxplot(df, title = "", base_size = 10)
}
\arguments{
\item{df}{dataframe or vector with continous or ordinal data}

\item{title}{Plot title}

\item{base_size}{numeric base font size}
}
\description{
Boxplot
}
\details{
uses ggplot
}
\examples{
vector<-generate_missing(rnorm(1000))
df<-generate_missing(mtcars[,1:2])
plot_boxplot(df=vector)
plot_boxplot(df=generate_missing(vector))
plot_boxplot(df=df)
}
\keyword{assumptions}
