% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_SEM.R
\name{plot_cfa}
\alias{plot_cfa}
\title{Plot cfa model}
\usage{
plot_cfa(model, ...)
}
\arguments{
\item{model}{lavaan object}

\item{...}{arguments passed to semPlot::semPaths}
}
\description{
Plot cfa model
}
\examples{
model='LATENT1=~X1+X2+X3
       LATENT2=~X4+X5+X6'
df<-lavaan::simulateData(model=model,model.type="cfa",
                             return.type="data.frame",sample.nobs=100)
df<-generate_missing(df)
fit<-lavaan::cfa(model,data=df,missing="ML")
plot_cfa(fit)
model='LATENT1=~X1+X2+X3+X4+X5+X6
       LATENT2=~X1+X2+X3+X4+X5+X6'
}
\keyword{SEM}
