% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_TRAIN_TEST.R
\name{plot_confusion}
\alias{plot_confusion}
\title{Plot confusion matrix}
\usage{
plot_confusion(observed, predicted, base_size = 10, title = "")
}
\arguments{
\item{observed}{vector of observed outcomes}

\item{predicted}{vector of predicted outcomes}

\item{base_size}{base font size}

\item{title}{plot title}
}
\description{
Plot confusion matrix
}
\examples{
plot_confusion(observed=c(1,2,3,1,2,3),predicted=c(1,2,3,1,2,3))
observed<-c(rep("male",10),rep("female",10),"male","male")
predicted<-c(rep("male",10),rep("female",10),"female","female")
plot_confusion(observed=observed,predicted=predicted)
}
\keyword{functions}
