% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_CORRELATION.R
\name{plot_corrplot}
\alias{plot_corrplot}
\title{Correlation matrix plots}
\usage{
plot_corrplot(mydata, title = "", base_size = 10, fill_limits = c(-1, 0, 1))
}
\arguments{
\item{mydata}{correlation matrix}

\item{title}{plot title}

\item{base_size}{base font size}

\item{fill_limits}{lower and upper limit for fill}
}
\description{
Correlation matrix plots
}
\examples{
plot_corrplot(stats::cor(mtcars),title="Correlation")
plot_corrplot(stats::cor(mtcars),base_size=20)
}
\keyword{correlation}
