% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPLORE_DESCRIPTIVES.R
\name{plot_crosstable}
\alias{plot_crosstable}
\title{Plot crosstables}
\usage{
plot_crosstable(
  df,
  factor_index,
  combinations = NULL,
  shape = 16,
  angle = 0,
  base_size = 10,
  title = ""
)
}
\arguments{
\item{df}{dataframe}

\item{factor_index}{index of factors}

\item{combinations}{index of comparisons}

\item{shape}{shape of points}

\item{angle}{angle of xaxis labels}

\item{base_size}{base font size}

\item{title}{plot title}
}
\description{
Plot crosstables
}
\examples{
combinations<-data.frame(index1=c("vs","am","gear"),index2=c("cyl","cyl","cyl"))
plot_crosstable(df=mtcars,factor_index=8:9)
plot_crosstable(df=mtcars,combinations=combinations)
}
\keyword{descriptives}
