% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPLORE_ASSUMPTIONS.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Histograms with density function}
\usage{
plot_histogram(
  df,
  bins = 30,
  title = "",
  base_size = 10,
  xlims = NULL,
  fill = "gray25",
  color = "gray50"
)
}
\arguments{
\item{df}{dataframe or vector with continous or ordinal data}

\item{bins}{number of bars to display}

\item{title}{plot title}

\item{base_size}{numeric base font size}

\item{xlims}{x axis limits}

\item{fill}{color of bar}

\item{color}{color of bar outline}
}
\description{
Histograms with density function
}
\details{
uses ggplot
}
\examples{
vector<-generate_missing(rnorm(1000))
df<-generate_missing(mtcars[,1:2])
plot_histogram(df=vector)
plot_histogram(df=df,xlims=c(0,10))
plot_histogram(df=df)
plot_multiplot(plotlist=plot_histogram(df=mtcars),cols=4)
}
\keyword{assumptions}
