% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_LOGISTIC_REGRESSION.R
\name{plot_logistic_model}
\alias{plot_logistic_model}
\title{Logistic model plot}
\usage{
plot_logistic_model(df, outcome = "outcome", title = "", base_size = 10)
}
\arguments{
\item{df}{dataframe with predictor and outcome outcome should be last}

\item{outcome}{name of outcome variable}

\item{title}{Character plot title}

\item{base_size}{base font size}
}
\description{
Logistic model plot
}
\examples{
df<-data.frame(outcome=c(rep(1,10),rep(0,10)),
               pd1=c(rep(1,11),rep(0,9)),
               pd2=c(rep(1,9),rep(0,11)),
               pc1=c(rnorm(10,mean=5),rnorm(10,mean=10)),
               pc2=c(rnorm(10,mean=5),rnorm(10,mean=20)))
plot_logistic_model(df=df,base_size=15)
}
\keyword{logistic}
\keyword{regression}
