% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPLORE_DESCRIPTIVES.R
\name{plot_mosaic}
\alias{plot_mosaic}
\title{Plot mosaic plots}
\usage{
plot_mosaic(df, factor_index, base_size = 10, title = "")
}
\arguments{
\item{df}{dataframe}

\item{factor_index}{index of factors}

\item{base_size}{base font size}

\item{title}{plot title}
}
\description{
Plot mosaic plots
}
\examples{
plot_mosaic(df=mtcars,factor_index=8:9)
plot_mosaic(df=mtcars,factor_index=9:10)
}
\keyword{descriptives}
