% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_PLOT.R
\name{plot_multiplot}
\alias{plot_multiplot}
\title{Multiple ggplot plots in one graph}
\usage{
plot_multiplot(..., plotlist = NULL, cols = 2, layout = NULL)
}
\arguments{
\item{...}{plot objects}

\item{plotlist}{a list of plots}

\item{cols}{number of columns in layout}

\item{layout}{a matrix specifying the layout. If present,'cols' is ignored}
}
\description{
Multiple ggplot plots in one graph
}
\examples{
p1<-ggplot(ChickWeight,aes(x=Time,y=weight,colour=Diet,group=Chick))+
           geom_line()+
           ggtitle("Growth curve for individual chicks")+
           theme_bw()
p2<-ggplot(ChickWeight,aes(x=Time,y=weight,colour=Diet))+
           geom_point(alpha=.3)+
           geom_smooth(alpha=.2,size=1,method="loess",formula="y~x")+
           ggtitle("Fitted growth curve per diet")+
           theme_bw()
p3<-ggplot(subset(ChickWeight,Time==21),aes(x=weight,colour=Diet))+
           geom_density()+
           ggtitle("Final weight, by diet")+theme_bw()
p4<-ggplot(subset(ChickWeight,Time==21),aes(x=weight,fill=Diet))+
           geom_histogram(colour="black",binwidth=50)+facet_grid(Diet~.)+
           ggtitle("Final weight, by diet")+theme_bw()
cars_plot<-plot_histogram(mtcars)
plot_multiplot(p1,p2,p3,p4,cols=2)
plot_multiplot(plotlist=plot_histogram(mtcars[,1:4]),cols=2)
plot_multiplot(plotlist=plot_histogram(mtcars),layout=matrix(1:4,ncol=2,byrow=TRUE))
plot_multiplot(plotlist=plot_scatterplot(mtcars[,1:4]),cols=2)
plot_multiplot(plotlist=cars_plot,layout=matrix(1:4,ncol=2,byrow=TRUE))
plot_multiplot(plotlist=cars_plot,cols=3)
}
\keyword{functions}
\keyword{plot}
