% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPLORE_ASSUMPTIONS.R
\name{plot_normality_diagnostics}
\alias{plot_normality_diagnostics}
\title{Normality plots}
\usage{
plot_normality_diagnostics(
  df,
  breaks = NULL,
  title = "",
  file = NULL,
  w = 10,
  h = 10
)
}
\arguments{
\item{df}{dataframe or vector with continous or ordinal data}

\item{breaks}{number of bars to display}

\item{title}{plot title}

\item{file}{output filename}

\item{w}{width of pdf file}

\item{h}{height of pdf file}
}
\description{
plot histogram density boxplot qq plot
}
\details{
uses plot base
}
\examples{
vector<-generate_missing(rnorm(1000))
df<-generate_missing(mtcars[,1:2])
plot_normality_diagnostics(df=vector,title="",file="rnorm",breaks=30)
plot_normality_diagnostics(df=vector,title="")
plot_normality_diagnostics(df=df,title="mtcars")
plot_normality_diagnostics(df=df,title="mtcars",file="rnorm")
}
\keyword{assumptions}
