% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_ANOVA_PLOT.R
\name{plot_oneway}
\alias{plot_oneway}
\title{Plot means with standard error for every level in a dataframe}
\usage{
plot_oneway(
  df,
  dv,
  iv,
  base_size = 20,
  type = "se",
  order_factor = TRUE,
  title = "",
  note = "",
  width = 60
)
}
\arguments{
\item{df}{dataframe}

\item{dv}{index of continous variables}

\item{iv}{index of factors}

\item{base_size}{base font size}

\item{type}{error bar type to display (1) "se" for standard error (2) "ci" for confidence interval (3) "sd" for standard deviation (4) "" for no error bar}

\item{order_factor}{if TRUE it will sort the categorical axis by the continous variable value}

\item{title}{plot title}

\item{note}{footnote}

\item{width}{wrap width for x title}
}
\description{
Plot means with standard error for every level in a dataframe
}
\examples{
nrows=1000
df<-data.frame(generate_factor(vector=LETTERS[1:5],nrows=nrows,ncols=10,type="random"),
               generate_data(nrows=nrows,ncols=5,type="normal"))
result<-plot_oneway(df=df,dv=11:15,iv=1:10)
plot_oneway(df=mtcars,dv=2,iv=9)
plot_oneway(df=mtcars,dv=2,iv=9:10)
plot_oneway(df=mtcars,dv=2:3,iv=10)
plot_oneway(df=mtcars,dv=2:3,iv=9:10)
plot_oneway(df=mtcars,dv=2:3,iv=9:10,type="se")
plot_oneway(df=mtcars,dv=2:3,iv=9:10,type="ci")
plot_oneway(df=mtcars,dv=2:3,iv=9:10,type="sd")
plot_oneway(df=mtcars,dv=2:3,iv=9:10,type="",order_factor=FALSE)
plot_oneway(df=mtcars,dv=2:3,iv=9:10,type="",order_factor=TRUE)
}
\keyword{ANOVA}
