% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_ANOVA_PLOT.R
\name{plot_oneway_diagnostics}
\alias{plot_oneway_diagnostics}
\title{Plot one way diagnostics}
\usage{
plot_oneway_diagnostics(df, dv, iv, base_size = 10)
}
\arguments{
\item{df}{dataframe}

\item{dv}{index of continous variables}

\item{iv}{index of factors}

\item{base_size}{base font size}
}
\description{
Plot one way diagnostics
}
\note{
Residuals vs Fitted should be equally spread horizontally otherwize the assumption of equality of variances is violated \cr
Normal QQ should show values in the diagonal otherwise the assumption of normality is violated
}
\examples{
nrows=1000
df<-data.frame(generate_factor(vector=LETTERS[1:5],nrows=nrows,ncols=10,type="random"),
               generate_data(nrows=nrows,ncols=5,type="normal"))
result<-plot_oneway_diagnostics(df=df,dv=11:15,iv=1:10)
plot_oneway_diagnostics(df=mtcars,dv=1:2,iv=9:10)
}
\keyword{ANOVA}
