% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPLORE_ASSUMPTIONS.R
\name{plot_outlier}
\alias{plot_outlier}
\title{Outlier graph using mean median and boxplot algorythms}
\usage{
plot_outlier(df, method = "mean", title = "", base_size = 10)
}
\arguments{
\item{df}{dataframe or vector with continous or ordinal data}

\item{method}{"mean" "median" "boxplot"}

\item{title}{plot title}

\item{base_size}{base font size}
}
\description{
Outlier graph using mean median and boxplot algorythms
}
\examples{
vector<-generate_missing(rnorm(1000))
df<-generate_missing(mtcars[,1:2])
plot_outlier(df=vector,method="mean",title="random vector")
plot_outlier(df=vector,method="median")
plot_outlier(df=vector,method="boxplot")
plot_outlier(df=df,method="mean",title="random vector")
plot_outlier(df=df,method="median")
plot_outlier(df=df,method="boxplot")
plot_multiplot(plotlist=plot_outlier(df=mtcars[,2:5],method="mean"),cols=2)
}
\author{
unknown
}
\keyword{assumptions}
