% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPLORE_ASSUMPTIONS.R
\name{plot_qq}
\alias{plot_qq}
\title{qq plots}
\usage{
plot_qq(df, title = "", base_size = 10)
}
\arguments{
\item{df}{dataframe or vector with continous or ordinal data}

\item{title}{plot title}

\item{base_size}{numeric base font size}
}
\description{
qq plots
}
\details{
uses ggplot
}
\examples{
vector<-generate_missing(rnorm(1000))
df<-generate_missing(mtcars[,1:2])
plot_qq(df=vector)
plot_qq(df=df)
plot_multiplot(plotlist=plot_qq(df=mtcars),cols=4)
}
\keyword{assumptions}
