% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPLORE_DESCRIPTIVES.R
\name{plot_response_frequencies}
\alias{plot_response_frequencies}
\title{Plot response frequencies}
\usage{
plot_response_frequencies(
  df,
  factor_index,
  base_size = 10,
  title = "",
  width = 100
)
}
\arguments{
\item{df}{dataframe}

\item{factor_index}{index of factors}

\item{base_size}{base font size}

\item{title}{plot title}

\item{width}{wrap width for x title}
}
\description{
Plot response frequencies
}
\examples{
plot_response_frequencies(df=mtcars,factor_index=1:10)
}
\keyword{descriptives}
