% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_LINEAR_REGRESSION.R
\name{plot_scatterplot}
\alias{plot_scatterplot}
\title{Plot plot_scatterplot}
\usage{
plot_scatterplot(
  df,
  method = lm,
  formula = y ~ x,
  base_size = 10,
  coord_equal = FALSE,
  all_orders = FALSE,
  title = "",
  combinations = NULL,
  string_aes = TRUE
)
}
\arguments{
\item{df}{dataframe if dataframe consists of 2 collumns the second collumn is the outcome and the first collumn is the predictor}

\item{method}{smoothing method, "auto", "lm", "glm", "gam", "loess" or a function, e.g. MASS::rlm or mgcv::gam, stats::lm, or stats::loess}

\item{formula}{formula used in smoothing function for geom_smooth}

\item{base_size}{base font size}

\item{coord_equal}{if TRUE axes maintain equal scale}

\item{all_orders}{if TRUE the order of combination is considered}

\item{title}{Plot title}

\item{combinations}{dataframe if not NULL user can provide a dataframe for variable combinations for x and y axis . First collumn represents x and second collumn represents y}

\item{string_aes}{if TRUE string_aes function is used for names}
}
\description{
Plot plot_scatterplot
}
\examples{
result<-plot_scatterplot(df=mtcars,title="",coord_equal=TRUE,base_size=10)
plot_multiplot(plotlist=result[1:12],cols=4)
plot_scatterplot(df=mtcars[,1:2],base_size=10,coord_equal=TRUE,all_orders=FALSE)
plot_scatterplot(df=mtcars[,1:2],base_size=10,coord_equal=FALSE,all_orders=FALSE)
plot_scatterplot(df=mtcars,base_size=10,coord_equal=TRUE,all_orders=FALSE,
                 combinations=data.frame(x=c("mpg","mpg","mpg"),
                                         y=c("cyl","hp","mpg")))
plot_scatterplot(df=mtcars,base_size=10,coord_equal=TRUE,all_orders=TRUE,
                 combinations=data.frame(x=c("mpg"),y=c("cyl")))
x<-rnorm(1000)
y<-x+rnorm(x,sd=.1)
plot_scatterplot(df=data.frame(x,y),title="Random Simulation",coord_equal=TRUE)
df<-data.frame(matrix(-.999,ncol=2,nrow=2))
correlation_martix<-as.matrix(df)
diag(correlation_martix)<-1
df<-generate_correlation_matrix(correlation_martix,nrows=1000)
plot_scatterplot(df,title="Simulation of -.999 Correlation",coord_equal=TRUE,base_size=20)
}
\keyword{regression}
