% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_EFA.R
\name{plot_scree}
\alias{plot_scree}
\title{Scree plot displaying the Kaiser and Jolife criteria for factor extraction}
\usage{
plot_scree(df, base_size = 15, title = "", color = c("#5F2C91", "#5E912C"))
}
\arguments{
\item{df}{dataframe}

\item{base_size}{base font size}

\item{title}{plot title}

\item{color}{color of line and point outline}
}
\description{
Scree plot displaying the Kaiser and Jolife criteria for factor extraction
}
\examples{
plot_scree(df=mtcars,title="",base_size=15)
}
\keyword{EFA}
