% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_TRAIN_TEST.R
\name{plot_separability}
\alias{plot_separability}
\title{Plot separability}
\usage{
plot_separability(observed, predicted, base_size = 10, title = "")
}
\arguments{
\item{observed}{vector of observed outcomes}

\item{predicted}{vector of predicted outcome probability}

\item{base_size}{base font size}

\item{title}{plot title}
}
\description{
Plot separability
}
\examples{
df1<-data.frame(matrix(.999,ncol=2,nrow=2))
correlation_martix<-as.matrix(df1)
diag(correlation_martix)<-1
df1<-generate_correlation_matrix(correlation_martix,nrows=1000)
df1$X1<-ifelse(abs(df1$X1)<1,0,1)
df1$X2<-abs(df1$X2)
df1$X2<-(df1$X2-min(df1$X2))/(max(df1$X2)-min(df1$X2))
plot_separability(observed=round(abs(df1$X1),0),predicted=abs(df1$X2))
}
\keyword{functions}
