% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPLORE_TIME_SERIES.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot timeseries}
\usage{
plot_ts(df, base_size = 10, ylab = "Count", title = "")
}
\arguments{
\item{df}{ts object}

\item{base_size}{base font size}

\item{ylab}{y label}

\item{title}{plot title}
}
\description{
Plot timeseries
}
\details{
returns plot
}
\examples{
ts_data<-ts(UKDriverDeaths,start=1969,end=1984,frequency=12)
result<-plot_ts(ts_data,title="UK driver deaths")
for(i in 1969:1984)
  result<-result+geom_vline(xintercept=i,color="blue",size=1,alpha=.5)
result
autoplot(stl(ts_data,s.window='periodic'))+
  theme_bw(base_size=10)+
  labs(title="UK driver deaths")
forecast::gglagplot(data.frame(ts_data),do.lines=FALSE,lags=100)+
  theme_bw(base_size=10)+labs(title="UK driver deaths",y="count")
}
\keyword{series}
\keyword{time}
