% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_STRINGS.R
\name{proper}
\alias{proper}
\title{Capitalize first character and lowercase the rest}
\usage{
proper(x)
}
\arguments{
\item{x}{Character}
}
\description{
Capitalize first character and lowercase the rest
}
\examples{
x<-generate_string(nchar=10,vector=LETTERS,vector_length=10)
proper(x)
}
\keyword{functions}
\keyword{strings}
