% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_TRAIN_TEST.R
\name{proportion_accurate}
\alias{proportion_accurate}
\title{Proportion overall accuracy of a confusion matrix}
\usage{
proportion_accurate(observed, predicted)
}
\arguments{
\item{observed}{vector of observed variables}

\item{predicted}{vector of predicted variables}
}
\description{
Proportion overall accuracy of a confusion matrix
}
\examples{
proportion_accurate(observed=c(1,2,3,4,5,10),predicted=c(1,2,3,4,5,11))
}
\keyword{functions}
