% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_KEYS.R
\name{questions_by_keys}
\alias{questions_by_keys}
\title{Convert key to index list}
\usage{
questions_by_keys(key)
}
\arguments{
\item{key}{a vector indicating the dimension of each question. The order of the elements in the key represents the order of the questions, the numeric values represent the dimension the question belongs to}
}
\description{
Convert key to index list
}
\examples{
key<-c(1,2,3,4,5,1,2,3,4,5)
questions_by_keys(key)
}
\keyword{dimensions}
\keyword{functions}
\keyword{keys}
