% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_KEYS.R
\name{questions_dimensions_dataframe}
\alias{questions_dimensions_dataframe}
\title{Question dimension table}
\usage{
questions_dimensions_dataframe(
  key,
  dimensions,
  elaborate_dimensions,
  questions
)
}
\arguments{
\item{key}{a vector indicating the dimension of each question. The order of the elements in the key represents the order of the questions, the numeric values represent the dimension the question belongs to}

\item{dimensions}{dimension names}

\item{elaborate_dimensions}{full dimension names}

\item{questions}{question names}
}
\description{
Return a dataframe with the order of the questions, their respective dimensions, and the description of the dimensions
}
\examples{
key<-c(1,2,3,4,5,1,2,3,4,5)
dimensions<-paste0("Dimension",1:10)
elaborate_dimensions<-paste0("Elaborated_Dimension",1:10)
questions<-paste0("Question",1:65)
questions_dimensions_dataframe(key,dimensions,elaborate_dimensions,questions)
}
\keyword{dimensions}
\keyword{functions}
\keyword{keys}
