% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_MATHEMATICAL.R
\name{rad2deg}
\alias{rad2deg}
\title{Convert radians to degrees}
\usage{
rad2deg(radians)
}
\arguments{
\item{radians}{radians}
}
\description{
Convert radians to degrees
}
\examples{
rad2deg(pi)
}
\keyword{correlation}
