% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_IRT_T.R
\name{rank3_to_triplets}
\alias{rank3_to_triplets}
\title{Convert thurstonian binary triplets to scale}
\usage{
rank3_to_triplets(mydata)
}
\arguments{
\item{mydata}{dataframe}
}
\description{
Convert thurstonian binary triplets to scale
}
\examples{
set.seed(12345)
mydata<-data.frame(i1=rnorm(10,mean=2,sd=.5),
                   i2=rnorm(10,mean=2,sd=.5),
                   i3=rnorm(10,mean=2,sd=.5),
                   i4=rnorm(10,mean=2,sd=.5),
                   i5=rnorm(10,mean=2,sd=.5),
                   i6=rnorm(10,mean=2,sd=.5))
result<-rank_to_binary(mydata[,1:3])
rank3_to_triplets(result)
}
\keyword{IRT}
\keyword{Thurstonian}
