% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_IRT_T.R
\name{rank_df_to_binary}
\alias{rank_df_to_binary}
\title{Convert scale to thurstonian binary with n items per block and n blocks}
\usage{
rank_df_to_binary(mydata, items, reverse = TRUE)
}
\arguments{
\item{mydata}{dataframe}

\item{items}{number of items in block}

\item{reverse}{if TRUE assumes that the highest value is first item in rank if FALSE the lowest value is the first item in rank}
}
\description{
Convert scale to thurstonian binary with n items per block and n blocks
}
\examples{
set.seed(12345)
mydata<-data.frame(i1=rnorm(10,mean=2,sd=.5),
                   i2=rnorm(10,mean=2,sd=.5),
                   i3=rnorm(10,mean=2,sd=.5),
                   i4=rnorm(10,mean=2,sd=.5),
                   i5=rnorm(10,mean=2,sd=.5),
                   i6=rnorm(10,mean=2,sd=.5))
rank_df_to_binary(mydata[,c("i1","i2","i3","i4")],4)
rank_df_to_binary(mydata,3)
}
\keyword{IRT}
\keyword{Thurstonian}
