% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_RELIABILITY.R
\name{raw_alpha}
\alias{raw_alpha}
\title{Raw alpha}
\usage{
raw_alpha(df)
}
\arguments{
\item{df}{dataframe with one dimension}
}
\description{
Raw alpha
}
\examples{
set.seed(12345)
df<-data.frame(matrix(.5,ncol=6,nrow=6))
correlation_martix<-as.matrix(df)
diag(correlation_martix)<-1
df<-round(generate_correlation_matrix(correlation_martix,nrows=1000),0)+5
psych::alpha(df)
raw_alpha(df=df)
}
\keyword{reliability}
