% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS.R
\name{rbind_all}
\alias{rbind_all}
\title{rbind dataframes or matrices with different lengths or collumn names}
\usage{
rbind_all(df1, df2)
}
\arguments{
\item{df1}{dataframe or matrix}

\item{df2}{dataframe or matrix}
}
\description{
rbind dataframes or matrices with different lengths or collumn names
}
\examples{
df1<-generate_correlation_matrix(n=10)
df2<-generate_correlation_matrix(n=10)
names(df2)[4]<-"X11"
rbind_all(df1=df1,df2=df2)
row.names(df1)<-21:30
rbind_all(df1=df1,df2=df2)
}
\keyword{functions}
