% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_RECODE.R
\name{recode_dataframe}
\alias{recode_dataframe}
\title{Takes a dataframe mapping response option number to response option text and recodes a numeric response vector to text response option for the entire dataframe}
\usage{
recode_dataframe(df, coding_map)
}
\arguments{
\item{df}{dataframe}

\item{coding_map}{dataframe containg map between coding and text}
}
\description{
Takes a dataframe mapping response option number to response option text and recodes a numeric response vector to text response option for the entire dataframe
}
\examples{
file<-"/opt/repo/rworking/projects/pcentral/data/aristotle-accountests-6-17.json"
pcentral<-rjson::fromJSON(file=file)
df<-json2dataframe(pcentral$itemresponses$biodata)
coding_map<-get_coding(pcentral$instr_descr_options$biodata)
names(df)[c(2,11,9,3,5,6,10,12,7)]<-names(coding_map)
recode_dataframe(df=df,coding_map=coding_map)
file<-"/opt/repo/rworking/projects/pcentral/data/aristotle-accountests-6-17.json"
pcentral<-rjson::fromJSON(file=file)
coding_map<-get_coding(pcentral$instr_descr_options$biodata)$Education
vector<-json2dataframe(pcentral$itemresponses$biodata)$education[1:50]
recode_dataframe(df=vector,coding_map=coding_map)
}
\keyword{functions}
