% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_TRAIN_TEST.R
\name{recode_scale_dummy}
\alias{recode_scale_dummy}
\title{Scale and dummy code}
\usage{
recode_scale_dummy(df, categories = 10)
}
\arguments{
\item{df}{dataframe}

\item{categories}{Numeric Number of unique values a vector must have to perform dummy coding}
}
\description{
Scales numeric variables between 0 and 1 and creates dummy coding for character and vector variables
}
\examples{
recode_scale_dummy(infert)
}
\keyword{functions}
