% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS.R
\name{remove_nc}
\alias{remove_nc}
\title{Replace remove non computable values}
\usage{
remove_nc(
  df,
  value = NA,
  remove_rows = FALSE,
  aggressive = FALSE,
  remove_cols = FALSE,
  remove_zero_variance = FALSE
)
}
\arguments{
\item{df}{dataframe}

\item{value}{replacement}

\item{remove_rows}{if TRUE it will remove rows with non computable values}

\item{aggressive}{if TRUE it will remove entire row if a single non computable value exists \cr
if FALSE it will remove row if all values are non computable}

\item{remove_cols}{if TRUE it will remove collumns with non computable values}

\item{remove_zero_variance}{if TRUE it will remove collumns with no variance}
}
\description{
Replace remove non computable values
}
\details{
Non computable values are NA, NAN, inf and empty cells.
}
\note{
This function internally replaces non computable values with the value choosen 
      the default value is NA. Then it removes rows and collumns with NA values or zero variance
}
\examples{
df<-mtcars
df[1,]<-as.numeric(NaN)
df[2,]<-as.numeric(Inf)
df[3,]<-as.numeric(-Inf)
df[4,]<-as.numeric(NA)
df[5,]<-""
remove_nc(df=df,value=NA)
cdf(remove_nc(df=df,value=NA))
df<-generate_missing(mtcars,missing=5)
remove_nc(df,remove_rows=TRUE,aggressive=FALSE)
remove_nc(df,remove_rows=TRUE,aggressive=TRUE)
df<-generate_missing(generate_correlation_matrix(nrows=5),missing=2)
df$X2<-NA
df$X3<-1
remove_nc(df,remove_cols=TRUE,remove_zero_variance=FALSE)
remove_nc(df,remove_cols=TRUE,remove_zero_variance=TRUE)
}
\keyword{functions}
