% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPLORE_ASSUMPTIONS.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers}
\usage{
remove_outliers(vector, probs = c(0.25, 0.75), na.rm = TRUE, ...)
}
\arguments{
\item{vector}{numeric}

\item{probs}{numeric vector with lowest and highest quantiles}

\item{na.rm}{if TRUE removes NA values}

\item{...}{arguments passed to quantile}
}
\description{
Remove outliers
}
\examples{
vector<-generate_missing(rnorm(1000))
df<-generate_missing(mtcars[,1:2])
remove_outliers(vector)
data.frame(sapply(df,remove_outliers))
}
\keyword{assumptions}
