% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS.R
\name{replace_na_with_previous}
\alias{replace_na_with_previous}
\title{Replace NA with the previous element in a vector}
\usage{
replace_na_with_previous(vector)
}
\arguments{
\item{vector}{Vector}
}
\description{
Replace NA with the previous element in a vector
}
\examples{
df1<-generate_missing(rnorm(10),missing=5)
df2<-generate_missing(rnorm(10),missing=5)
df3<-generate_missing(rnorm(10),missing=5)
df4<-generate_missing(rnorm(10),missing=5)
df5<-generate_missing(rnorm(10),missing=5)
df<-data.frame(df1,df2,df3,df4,df5)
row.names(df)<-paste0("A",row.names(df))
replace_na_with_previous(df1)
df[]<-lapply(df,replace_na_with_previous)
}
\keyword{functions}
