% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_CORRELATION.R
\name{report_choric_serial}
\alias{report_choric_serial}
\title{Report polychoric tetrachoric polyserial biserial correlation}
\usage{
report_choric_serial(
  x,
  y = NULL,
  file = NULL,
  w = 10,
  h = 10,
  type = "tetrachoric",
  ...
)
}
\arguments{
\item{x}{The input may be in one of four forms:\cr
a) a data frame or matrix of dichotmous data (e.g., the lsat6 from the bock data set) or discrete numerical (i.e., not too many levels, e.g., the big 5 data set, bfi) for polychoric, or continuous for the case of biserial and polyserial\cr
b) a 2 x 2 table of cell counts or cell frequencies (for tetrachoric) or an n x m table of cell counts (for both tetrachoric and polychoric)\cr
c) a vector with elements corresponding to the four cell frequencies (for tetrachoric)\cr
d) a vector with elements of the two marginal frequencies (row and column) and the comorbidity (for tetrachoric)\cr}

\item{y}{matrix or dataframe of discrete scores. In the case of tetrachoric, these should be dichotomous, for polychoric not too many levels, for biserial they should be discrete (e.g., item responses) with not too many (<10?) categories}

\item{file}{output filename}

\item{w}{width of pdf file}

\item{h}{height of pdf file}

\item{type}{"tetrachoric" "polychoric" "polyserial" "biserial"}

\item{...}{arguments passed to psych::polychoric}
}
\description{
Report polychoric tetrachoric polyserial biserial correlation
}
\examples{
report_choric_serial(generate_data(min=0,max=1,type="uniform"),
              type="tetrachoric",file="tetrachoric")
report_choric_serial(generate_data(min=1,max=5,type="uniform"),
              type="polychoric")
report_choric_serial(x=psych::lsat6,y=psych::lsat6,
                     type="polyserial",file="polyserial")
report_choric_serial(x=psych::lsat6,y=psych::lsat6,
                     type="biserial",file="biserial")
}
\keyword{correlation}
