% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_EXCEL.R
\name{report_dataframe}
\alias{report_dataframe}
\title{Write matrix or dataframe to excel sheet}
\usage{
report_dataframe(df, file = NULL, type = "critical_value", ...)
}
\arguments{
\item{df}{dataframe or matrix}

\item{file}{output filename of excel file}

\item{type}{"critical_value" "matrix"}

\item{...}{arguments passed to excel_critical_value or to excel_matrix}
}
\description{
Usefull for generic data where conditional formating of a spesific collumn is required
}
\examples{
comment<-list(mpg="Miles/(US) gallon",
              cyl="Number of cylinders",
              disp="Displacement (cu.in.)",
              hp="Gross horsepower",
              drat="Rear axle ratio",
              wt="Weight (1000 lbs)",
              qsec="1/4 mile time",
              vs="Engine (0=V-shaped,1=straight)",
              am="Transmission (0=automatic,1=manual)",
              gear="Number of forward gears",
              carb="Number of carburetors")
report_dataframe(mtcars,sheet="report",file="mtcars",comment=comment,numFmt="#0.00",
                 critical=list(am="<0.05"))
report_dataframe(mtcars,sheet="report",file=NULL,comment=comment,numFmt="#0.00",
                 critical=list(am="<0.05"))
}
\keyword{functions}
