% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_HLR.R
\name{report_hlr}
\alias{report_hlr}
\title{Report HLR}
\usage{
report_hlr(
  df,
  corlist,
  factorlist,
  predictor,
  random_effect,
  file = NULL,
  sheet = "report"
)
}
\arguments{
\item{df}{dataframe}

\item{corlist}{Numeric outcome index}

\item{factorlist}{Numeric predictor index}

\item{predictor}{Character predictor name}

\item{random_effect}{Character random effect name}

\item{file}{Character file}

\item{sheet}{Character sheet}
}
\description{
Report HLR
}
\examples{
# report_hlr(df=infert,corlist=8,factorlist=1,
#           predictor="case",random_effect="case")
}
