% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_LDA.R
\name{report_lda}
\alias{report_lda}
\title{Report for MASS::lda}
\usage{
report_lda(model, file = NULL, w = 10, h = 10, base_size = 10, title = "")
}
\arguments{
\item{model}{object from MASS::lda}

\item{file}{output filename}

\item{w}{width of pdf file}

\item{h}{height of pdf file}

\item{base_size}{base font size}

\item{title}{plot title}
}
\description{
Report for MASS::lda
}
\examples{
model<-MASS::lda(case~.,data=infert)
report_lda(model=model)
report_lda(model=model,file="lda")
model<-MASS::lda(Species~.,data=iris)
report_lda(model=model,file="lda")
}
\keyword{ML}
