% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_LOGISTIC_REGRESSION.R
\name{report_logistic}
\alias{report_logistic}
\title{Report logistic regression}
\usage{
report_logistic(
  model,
  validation_data = NULL,
  file = NULL,
  title = "",
  w = 10,
  h = 10,
  base_size = 10,
  fast = FALSE
)
}
\arguments{
\item{model}{object glm}

\item{validation_data}{validation data}

\item{file}{output filename}

\item{title}{plot title}

\item{w}{width of pdf file. Relevant only when file string is not empty}

\item{h}{height of pdf file. Relevant only when file string is not empty}

\item{base_size}{base font size}

\item{fast}{if TRUE it will not output individual scores and residuals}
}
\description{
Report logistic regression
}
\note{
(1) Problematic values for standardized residuals > +-1.96 \cr
Standardized residuals are residuals divided by an estimated standard deviation and they can be interpreted as z scores in that: \cr
95% of z-scores should lie between -1.96 and +1.96 \cr
99% of z-scores should lie between -2.58 and +2.58 \cr 
99.99% of z-scores should lie between -3.29 and +3.29 \cr
(2) Problematic values for dfBeta >=1 \cr
dfBeta estimates coefficients if the respective case is removed from the dataset \cr
(3) Problematic values for Hat values (leverage) 2 or 3 times the average (k+1/n) \cr
Hat values (leverage),gauge the influence of the observed value of the outcome variable over the predicted values \cr
The average leverage value is defined as (k+1)/n, k=number of predictors, n=number of participants. Leverage values lie between 0 (no influence) and 1 (complete influence over prediction) \cr
If no cases exert undue influence over the model then all leverage values should be close to (k+1)/n \cr
Hoaglin and Welsch (1978) recommends investigating cases with values greater than twice the average (2(k+1)/n) \cr
Stevens (2002) recommends investigating cases with values greater than three times the average (3(k+1)/n) \cr
(4) Problematic values for VIFs > 10 \cr
ASSUMPTIONS  \cr
(1) Linearity between continous predictors and the logit (test wether the interaction term between the predictor and its log transformation is significant) \cr
(2) Independence of errors \cr
(3) No multicolinearity \cr
}
\examples{
modelcategoricalpredictor0<-glm(case~education,data=infert,family=binomial)
modelcategoricalpredictor1<-glm(case~education,data=infert,family=gaussian)
#modelcategoricalpredictor2<-glm(case~education,data=infert,family=Gamma)
#modelcategoricalpredictor3<-glm(case~education,data=infert,family=inverse.gaussian)
modelcategoricalpredictor4<-glm(case~education,data=infert,family=poisson)
modelcategoricalpredictor5<-glm(case~education,data=infert,family=quasi)
modelcategoricalpredictor6<-glm(case~education,data=infert,family=quasibinomial)
modelcategoricalpredictor7<-glm(case~education,data=infert,family=quasipoisson)
modelcontinuouspredictor0<-glm(case~stratum,data=infert,family=binomial)
modeltwopredictors0<-glm(case~education+stratum,data=infert,family=binomial)
modeltwopredictors1<-glm(case~education+stratum,data=infert,family=gaussian)
#modeltwopredictors2<-glm(case~education+stratum,data=infert,family=Gamma)
#modeltwopredictors3<-glm(case~education+stratum,data=infert,family=inverse.gaussian)
modeltwopredictors4<-glm(case~education+stratum,data=infert,family=poisson)
modeltwopredictors5<-glm(case~education+stratum,data=infert,family=quasi)
modeltwopredictors6<-glm(case~education+stratum,data=infert,family=quasibinomial)
modeltwopredictors7<-glm(case~education+stratum,data=infert,family=quasipoisson)
report_logistic(model=modelcategoricalpredictor0)
report_logistic(model=modelcategoricalpredictor1)
#report_logistic(model=modelcategoricalpredictor2)
#report_logistic(model=modelcategoricalpredictor3)
report_logistic(model=modelcategoricalpredictor4)
report_logistic(model=modelcategoricalpredictor5)
report_logistic(model=modelcategoricalpredictor6)
report_logistic(model=modelcategoricalpredictor7)
report_logistic(model=modelcontinuouspredictor0)
report_logistic(model=modeltwopredictors0)
report_logistic(model=modelcategoricalpredictor0,
                file="logistic_categorical_predictor",
                validation_data=infert)
report_logistic(model=modelcontinuouspredictor0,
                file="logistic_continuous_predictor",
                validation_data=infert)
report_logistic(model=modeltwopredictors0,
                file="logistic_two_predictors",
                validation_data=infert[1:10,])
}
\keyword{logistic}
\keyword{regression}
