% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPLORE_ASSUMPTIONS.R
\name{report_normality_tests}
\alias{report_normality_tests}
\title{Normality tests}
\usage{
report_normality_tests(df, file = NULL)
}
\arguments{
\item{df}{dataframe with continous or ordinal data}

\item{file}{output filename}
}
\description{
Shapiro-Wilk Anderson-Darling Cramer-von-Mises Shapiro-Francia Jarque-Bera Kolmogorov-Smirnov Lilliefors Pearson X2
}
\details{
returns xlsx file
}
\examples{
vector<-generate_missing(rnorm(1000))
df<-generate_missing(mtcars[,1:2])
report_normality_tests(df=df)
report_normality_tests(df=vector,file="normality_tests")
}
\keyword{assumptions}
