% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_PLOT.R
\name{report_pdf}
\alias{report_pdf}
\title{Report pdf}
\usage{
report_pdf(
  ...,
  plotlist = NULL,
  file = NULL,
  title = NULL,
  w = 10,
  h = 10,
  print_plot = TRUE
)
}
\arguments{
\item{...}{plot objects}

\item{plotlist}{list of plot objects}

\item{file}{output filename}

\item{title}{output filename}

\item{w}{width of pdf file}

\item{h}{height of pdf file}

\item{print_plot}{if TRUE it prints plot on graphics device}
}
\description{
Report pdf
}
\examples{
p1<-ggplot(ChickWeight,aes(x=Time,y=weight,colour=Diet,group=Chick))+
           geom_line()+
           ggtitle("Growth curve for individual chicks")+
           theme_bw()
p2<-ggplot(ChickWeight,aes(x=Time,y=weight,colour=Diet))+
           geom_point(alpha=.3)+
           geom_smooth(alpha=.2,size=1,method="loess",formula="y~x")+
           ggtitle("Fitted growth curve per diet")+theme_bw()
cars_plot_multiplot<-plot_multiplot(plotlist=plot_histogram(mtcars[,1:4]),cols=2)
cars_plot_base<-plot_normality_diagnostics(mtcars)
report_pdf(p1,p2,print_plot=TRUE)
report_pdf(p1,p2,file="report",print_plot=FALSE)
report_pdf(plotlist=cars_plot_multiplot,print_plot=TRUE)
report_pdf(plotlist=cars_plot_multiplot,file="report",print_plot=FALSE)
report_pdf(plotlist=cars_plot_base,print_plot=TRUE)
report_pdf(plotlist=cars_plot_base,file="report",print_plot=FALSE)
}
\keyword{functions}
\keyword{plot}
