% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_MEANS.R
\name{report_ttests}
\alias{report_ttests}
\title{T test}
\usage{
report_ttests(df, dv, iv, file = NULL, ...)
}
\arguments{
\item{df}{dataframe}

\item{dv}{index of continous variables}

\item{iv}{index of factors}

\item{file}{output filename}

\item{...}{
  Arguments passed on to \code{\link[stats:t.test]{stats::t.test}}
  \describe{
    \item{\code{x}}{a (non-empty) numeric vector of data values.}
  }}
}
\description{
T test
}
\examples{
report_ttests(df=mtcars,dv=2,iv=9:10)
report_ttests(df=mtcars,dv=2:3,iv=9)
report_ttests(df=mtcars,dv=2:3,iv=9:10,alternative="two.sided")
report_ttests(df=mtcars,dv=2:7,iv=9:10,alternative="less")
report_ttests(df=mtcars,dv=2:7,iv=9:10,alternative="greater")
report_ttests(df=mtcars,dv=1:7,iv=8:10,var.equal=TRUE)
report_ttests(df=mtcars,dv=1:7,iv=8:10,var.equal=TRUE,file="ttest")
}
\keyword{ANOVA}
