% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_MEANS.R
\name{report_wtests}
\alias{report_wtests}
\title{Wilcoxon test}
\usage{
report_wtests(df, dv, iv, file = NULL, ...)
}
\arguments{
\item{df}{dataframe}

\item{dv}{index of continous variables}

\item{iv}{index of factors}

\item{file}{output filename}

\item{...}{
  Arguments passed on to \code{\link[stats:wilcox.test]{stats::wilcox.test}}
  \describe{
    \item{\code{x}}{numeric vector of data values.  Non-finite (e.g., infinite or
    missing) values will be omitted.}
  }}
}
\description{
Wilcoxon test
}
\examples{
report_wtests(df=mtcars,dv=2,iv=9)
report_wtests(df=mtcars,dv=2,iv=9:10)
report_wtests(df=mtcars,dv=2:3,iv=9)
report_wtests(df=mtcars,dv=2:3,iv=9:10,alternative="two.sided")
report_wtests(df=mtcars,dv=2:7,iv=9:10,alternative="less")
report_wtests(df=mtcars,dv=2:7,iv=9:10,alternative="greater")
report_wtests(df=mtcars,dv=1:7,iv=8:10,var.equal=TRUE)
report_wtests(df=mtcars,dv=1:7,iv=8:10,var.equal=TRUE,file="wilcoxontest")
}
\keyword{ANOVA}
