% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EXPLORE_DESCRIPTIVES.R
\name{response_frequency}
\alias{response_frequency}
\title{Response frequencies}
\usage{
response_frequency(
  df,
  max = 10,
  uniqueitems = NULL,
  type = "percent",
  file = NULL
)
}
\arguments{
\item{df}{dataframe}

\item{max}{maximum score}

\item{uniqueitems}{number of unique items}

\item{type}{"frequency" "proportion" "percent" "all"}

\item{file}{output filename}
}
\description{
returns count proportion percent
}
\details{
returns dataframe
}
\examples{
response_frequency(mtcars[,c("gear","carb")],uniqueitems=1:8,type="frequency")
response_frequency(mtcars[,c("gear")],uniqueitems=1:8,type="proportion")
response_frequency(mtcars[,c("gear","carb")],uniqueitems=1:8,type="percent")
response_frequency(mtcars[,c("gear","carb")],uniqueitems=1:8,type="all")
response_frequency(mtcars[,c("gear","carb")],uniqueitems=1:8,type="all",
                   file="descriptives")
}
\keyword{descriptives}
