% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_RELIABILITY.R
\name{shrout}
\alias{shrout}
\title{Shrout reliability}
\usage{
shrout(sperson, spersonitem, stime, spersontime, serror, m, k)
}
\arguments{
\item{sperson}{variance component of participant}

\item{spersonitem}{variance component of participant by item interaction}

\item{stime}{variance component of time}

\item{spersontime}{variance component of participant by time interaction}

\item{serror}{variance component of error}

\item{m}{m item reports}

\item{k}{k time points}
}
\description{
Shrout reliability
}
\examples{
design<-expand.grid(time=1:3,item=1:2,person=1:10)
design<-change_data_type(design,type="factor")
design$response<-rnorm(30,0,0.1)
model<-mixlm::lm(response~r(time)*r(person)+r(item)*r(person),data=design)
result<-extract_components(model)
vc<-result$components
shrout(sperson=vc[2,3],spersonitem=vc[5,3],stime=vc[1,3],
       spersontime=vc[4,3],serror=vc[6,3],3,3)
}
\keyword{reliability}
