% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_SEM.R
\name{simulate_cfa_fit}
\alias{simulate_cfa_fit}
\title{Simulate CFA from coefficients}
\usage{
simulate_cfa_fit(
  model_sim = NULL,
  model = NULL,
  df = NULL,
  minnobs = 50,
  maxnobs = 1000,
  stepping = 10,
  file = NULL,
  w = 10,
  h = 10
)
}
\arguments{
\item{model_sim}{lavaan model spesification with defined coefficients}

\item{model}{lavaan model spesification with free coefficients}

\item{df}{dataframe}

\item{minnobs}{start sample size}

\item{maxnobs}{end sample size}

\item{stepping}{stepping}

\item{file}{output filename}

\item{w}{width of pdf file}

\item{h}{height of pdf file}
}
\description{
Simulates cfa from coefficients
             Simulates cfa from correlations of obeserved data
             Returns fit indices for predefined set of sample sizes
}
\examples{
model_sim='LATENT=~1*X1+0.5*X2+1.5*X3+1.5*X4+X5'
model='LATENT=~X1+X2+X3+X4+X5'
df<-lavaan::simulateData(model=model_sim,model.type="cfa",
                         return.type="data.frame",sample.nobs=1000)
# simulate_cfa_fit(model_sim=model_sim,model=model,
#                  minnobs=50,maxnobs=1000,stepping=100,file="report")
# simulate_cfa_fit(model=model,df=df,
#                  minnobs=50,maxnobs=1000,stepping=100,file="report")
}
\keyword{SEM}
