% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_GENERATEDATA.R
\name{simulate_correlation_from_sample}
\alias{simulate_correlation_from_sample}
\title{Generate a dataframe that produces the same correlation matrix as the input dataframe}
\usage{
simulate_correlation_from_sample(cordata, nrows = 10)
}
\arguments{
\item{cordata}{dataframe}

\item{nrows}{number of rows to generate}
}
\description{
Generate a dataframe that produces the same correlation matrix as the input dataframe
}
\examples{
correlation_matrix<-generate_correlation_matrix()
stats::cor(correlation_matrix)
simulate_correlation_from_sample(correlation_matrix,nrows=1000)
stats::cor(simulate_correlation_from_sample(correlation_matrix,nrows=1000))
}
\keyword{data}
\keyword{functions}
\keyword{generate}
