% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_STRINGS.R
\name{split_str_df}
\alias{split_str_df}
\title{Split string in dataframe}
\usage{
split_str_df(df, split = "/", type = "row", index, ...)
}
\arguments{
\item{df}{dataframe}

\item{split}{Separation character}

\item{type}{"row" "collumn" 
if "row" it will split the string of row names and it will display it on seperate collumns 
if "collumn" it will split the string of a spesified collumn and it will display it on separate collumns}

\item{index}{Numeric index of collumn to split. This is only relevant if type="collumn"}

\item{...}{arguments passed to split_str}
}
\description{
Split string in dataframe
}
\examples{
df<-generate_correlation_matrix()
string<-paste0(1:nrow(df),"/",
               generate_string(nchar=2,vector_length=nrow(df)),"/",
               generate_string(nchar=2,vector_length=nrow(df)),"/",
               generate_string(nchar=2,vector_length=nrow(df)))
row.names(df)<-string
split_str_df(df,split="/",type="row")
df[,1]<-string
split_str_df(df,split="/",type="collumn",index=1)
}
\keyword{functions}
\keyword{strings}
