% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_STRINGS.R
\name{string_aes}
\alias{string_aes}
\title{Adjust string aesthetics}
\usage{
string_aes(
  vector,
  characterlist = c(".", "_", "-", ",", "$", "<p>", "</p>", "<br>", "<br/>", "<B>",
    "</B>", "<BR/>", "|", "/", "&nbsp"),
  proper = TRUE
)
}
\arguments{
\item{vector}{Vector}

\item{characterlist}{List the list of characters to treat as separating characters}

\item{proper}{Logical TRUE capitalizes the first letter in sentense format}
}
\description{
Treats spesific characters such as ".", as separating characters and separates strings with space. Trims leading and trailing spaces and capitalizes the first letter of the string and lowers the rest.
}
\examples{
vector<-c("TES.T","TES<p>T","TES&nbspT")
string_aes(vector=vector)
string_aes(vector=vector,proper=FALSE)
string_aes(vector=vector,proper=TRUE)
}
\keyword{functions}
\keyword{strings}
