% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_STRINGS.R
\name{sub_str}
\alias{sub_str}
\title{Return n characters from left or right}
\usage{
sub_str(x, n = 2, type)
}
\arguments{
\item{x}{Character}

\item{n}{Number of characters to return}

\item{type}{"right" "left"}
}
\description{
Return n characters from left or right
}
\examples{
sub_str("12345",n=2,type="right")
sub_str("12345",n=2,type="left")
}
\keyword{functions}
\keyword{strings}
