% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_RECODE.R
\name{swap}
\alias{swap}
\title{Reverse a numeric vector}
\usage{
swap(vector)
}
\arguments{
\item{vector}{numeric}
}
\description{
Reverse a numeric vector
}
\examples{
swap(c(1:10,1,2,3))
}
\keyword{functions}
