% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_MATRIX.R
\name{symmetric_matrix}
\alias{symmetric_matrix}
\title{Symmetric Matrix}
\usage{
symmetric_matrix(matrix, duplicate = "lower", diagonal = NULL)
}
\arguments{
\item{matrix}{matrix}

\item{duplicate}{"upper" duplicates upper triangle "lower" duplicates lower triangle}

\item{diagonal}{diagonal values}
}
\description{
Symmetric Matrix
}
\examples{
m_lower<-matrix_triangle(matrix(1:9,nrow=3,ncol=3),type="lower",diagonal=NA)
m_upper<-matrix_triangle(matrix(11:19,nrow=3,ncol=3),type="upper",diagonal=NA)
symmetric_matrix(matrix=m_lower,duplicate="lower",diagonal=NA)
symmetric_matrix(matrix=m_upper,duplicate="upper",diagonal=NA)
}
\keyword{functions}
\keyword{matrix}
