% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_STRINGS.R
\name{trim_df}
\alias{trim_df}
\title{Trim whitespace in dataframe}
\usage{
trim_df(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Trim whitespace in dataframe
}
\examples{
string<-data.frame(str1=rep(paste0(sample(c(LETTERS,rep(" ",10))),collapse=""),10),
                   str2=rep(paste0(sample(c(LETTERS,rep(" ",10))),collapse=""),10),
                   num1=rnorm(10),
                   stringsAsFactors=FALSE)
trim_df(string)
}
\keyword{functions}
\keyword{strings}
