% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_PLOT.R
\name{wrapper}
\alias{wrapper}
\title{Wrap string}
\usage{
wrapper(x, ...)
}
\arguments{
\item{x}{title}

\item{...}{arguments passed to strwrap}
}
\description{
Wrap string
}
\examples{
wrapper(rep("sting",50),30)
}
\keyword{functions}
\keyword{plot}
