% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_ENVIRONMENT.R
\name{write_txt}
\alias{write_txt}
\title{Log console in file}
\usage{
write_txt(input, file = NULL)
}
\arguments{
\item{input}{Script to log in log file}

\item{file}{Filename of log}
}
\description{
Logs console in file and then displays log in console
}
\examples{
write_txt(mtcars)
write_txt(mtcars,file="mtcars")
}
\keyword{functions}
