% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP.R
\name{clear_stopwords}
\alias{clear_stopwords}
\title{Remove stopwods}
\usage{
clear_stopwords(
  text,
  stopwords = unique(c(tm::stopwords("SMART"), corpus::stopwords_en))
)
}
\arguments{
\item{text}{character vector}

\item{stopwords}{character words to remove}
}
\description{
Remove stopwods
}
\examples{
text1<-"word_one word_two word_three"
text2<-"word_three word_four word_six"
text3<-"All the Lorem Ipsum generators on the Internet tend to repeat predefined 
chunks as necessary, making this the first true generator on the Internet."
text4<-"It uses a dictionary of over 200 Latin words, combined with a handful of 
model sentence structures, to generate Lorem Ipsum which looks reasonable."
text5<-"The generated Lorem Ipsum is therefore always free from repetition, 
injected humour, or non-characteristic words etc."
text<-c(text1,text2,text3,text4,text5)
clear_stopwords(text)
}
\keyword{NLP}
