% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_ANOVA.R
\name{compute_kruskal_wallis_test}
\alias{compute_kruskal_wallis_test}
\title{Kruskal Wallis test}
\usage{
compute_kruskal_wallis_test(formula, df)
}
\arguments{
\item{formula}{one way formula in form of y~x. It will ignore more complex formulas}

\item{df}{dataframe
eta squared ranges between 0 and 1 \cr
epsilon squared ranges between 0 and 1 \cr
eta squared multiplied by 100 indicates the percentage of variance in the dependent variable explained by the independent variable \cr}
}
\description{
Kruskal Wallis test
}
\examples{
form<-formula(bp_before~agegrp)
kruskal.test(formula=form,data=df_blood_pressure)
rcompanion::epsilonSquared(x=df_blood_pressure$bp_before,
                           g=df_blood_pressure$agegrp,
                           group="row",
                           ci=TRUE,
                           conf=0.95,
                           type="perc",
                           R=1000,
                           digits=3)
rstatix::kruskal_effsize(df_blood_pressure,form,ci=TRUE,conf.level=0.95,ci.type="perc",nboot=100)
compute_kruskal_wallis_test(formula=form,df=df_blood_pressure)
}
\keyword{ANOVA}
