% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLM_ANOVA.R
\name{compute_posthoc}
\alias{compute_posthoc}
\title{Games Howell Tukey post hoc tests}
\usage{
compute_posthoc(y, x, method = c("games-howell", "tukey"))
}
\arguments{
\item{y}{Vector continous variable}

\item{x}{Vector factor}

\item{method}{Character "games-howell" "tukey" or c("games-howell","tukey")}
}
\description{
Based on http://www.psych.yorku.ca/cribbie/6130/games_howell.R
}
\examples{
result<-compute_posthoc(mtcars[,6],mtcars[,10])
}
\keyword{ANOVA}
