% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_TRAIN_TEST.R
\name{confusion}
\alias{confusion}
\title{Create a confusion matrix from observed and expected vectors}
\usage{
confusion(observed, predicted)
}
\arguments{
\item{observed}{vector of observed variables}

\item{predicted}{vector of predicted variables}
}
\description{
Create a confusion matrix from observed and expected vectors
}
\examples{
confusion(observed=c(1,2,3,4,5,10),predicted=c(1,2,3,4,5,11))
confusion(observed=c(1,2,2,2,2),predicted=c(1,1,2,2,2))
}
\keyword{functions}
