% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_TRAIN_TEST.R
\name{confusion_matrix_percent}
\alias{confusion_matrix_percent}
\title{Confusion matrix with row and column percent}
\usage{
confusion_matrix_percent(observed, predicted)
}
\arguments{
\item{observed}{vector of observed variables}

\item{predicted}{vector of predicted variables}
}
\description{
Confusion matrix with row and column percent
}
\note{
Total measures - Accuracy: (TP+TN)/total\cr
Total measures - Prevalence: (TP+FN)/total\cr
Total measures - Proportion Incorrectly Classified: (FN+FP)/total\cr
Horizontal measures - True Positive Rate - Sensitivity: TP/(TP+FN)\cr
Horizontal measures - True Negative Rate - Specificity: TN/(FP+TN)\cr
Horizontal measures - False Negative Rate - Miss Rate: FN/(TP+FN)\cr
Horizontal measures - False Positive Rate - Fall-out: FP/(FP+TN)\cr
Vertical measures - Positive Predictive value - Precision: TP/(TP+FP)\cr
Vertical measures - Negative Predictive value: TN/(FN+TN)\cr
Vertical measures - False Omission Rate: FN/(FN+TN)\cr
Vertical measures - False Discovery Rate: FP/(TP+FP)\cr
}
\examples{
observed<-factor(round(rnorm(10000,m=10,sd=1)))
predicted<-factor(round(rnorm(10000,m=10,sd=1)))
confusion_matrix_percent(observed,predicted)
}
\keyword{functions}
