% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{df_automotive_data}
\alias{df_automotive_data}
\title{Automotive Data}
\format{
A data frame with 38 rows and 26 variables:
\describe{
  \item{engine.location}{Location of the engine (e.g., front, rear)}
  \item{wheel.base}{Wheelbase of the vehicle in inches}
  \item{length}{Length of the vehicle in inches}
  \item{width}{Width of the vehicle in inches}
  \item{height}{Height of the vehicle in inches}
  \item{curb.weight}{Curb weight of the vehicle in pounds}
  \item{engine.type}{Type of engine (e.g., dohc, ohcv, ohc, l)}
  \item{num.of.cylinders}{Number of cylinders in the engine}
  \item{engine.size}{Size of the engine in cubic inches}
  \item{fuel.system}{Fuel system used (e.g., mpfi, 2bbl, mfi, 1bbl)}
  \item{bore}{Diameter of the cylinders in the engine}
  \item{stroke}{Stroke length of the engine}
  \item{compression.ratio}{Compression ratio of the engine}
  \item{horsepower}{Horsepower generated by the engine}
  \item{peak.rpm}{Peak RPM of the engine}
  \item{city.mpg}{Miles per gallon in the city}
  \item{highway.mpg}{Miles per gallon on the highway}
  \item{price}{Price of the vehicle}
}
}
\source{
Downloaded from Kaggle.com by the user Ramakrishnan Srinivasan. 
see \url{https://www.kaggle.com/toramky/automobile-dataset}
}
\usage{
df_automotive_data
}
\description{
This data set contains various automotive information including 
engine location, dimensions, weight, engine type, number of cylinders, 
and other specifications.
}
\keyword{datasets}
