% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA.R
\docType{data}
\name{df_crop_yield}
\alias{df_crop_yield}
\title{Automotive Data}
\format{
A data frame with 38 rows and 26 variables:
\describe{
  \item{engine.location}{Location of the engine (e.g., front, rear)}
  \item{wheel.base}{Wheelbase of the vehicle in inches}
  \item{length}{Length of the vehicle in inches}
  \item{width}{Width of the vehicle in inches}
  \item{height}{Height of the vehicle in inches}
  \item{curb.weight}{Curb weight of the vehicle in pounds}
  \item{engine.type}{Type of engine (e.g., dohc, ohcv, ohc, l)}
  \item{num.of.cylinders}{Number of cylinders in the engine}
  \item{engine.size}{Size of the engine in cubic inches}
  \item{fuel.system}{Fuel system used (e.g., mpfi, 2bbl, mfi, 1bbl)}
  \item{bore}{Diameter of the cylinders in the engine}
  \item{stroke}{Stroke length of the engine}
  \item{compression.ratio}{Compression ratio of the engine}
  \item{horsepower}{Horsepower generated by the engine}
  \item{peak.rpm}{Peak RPM of the engine}
  \item{city.mpg}{Miles per gallon in the city}
  \item{highway.mpg}{Miles per gallon on the highway}
  \item{price}{Price of the vehicle}
}

A data frame with 30 rows and 5 variables:
\describe{
  \item{patient}{Unique identifier for each patient}
  \item{sex}{Sex of the patient (e.g., Male, Female)}
  \item{agegrp}{Age group of the patient (e.g., 30-45, 46-59)}
  \item{bp_before}{Blood pressure reading before the intervention}
  \item{bp_after}{Blood pressure reading after the intervention}
}

A data frame with [number of rows] rows and [number of variables] variables:
\describe{
  \item{variable1}{Description of variable1}
  \item{variable2}{Description of variable2}
  \item{variable3}{Description of variable3}
  \item{variable4}{Description of variable4}
  \item{variable5}{Description of variable5}
  \item{variable6}{Description of variable6}
}
}
\source{
Downloaded from Kaggle.com by the user Ramakrishnan Srinivasan. \url{https://www.kaggle.com/toramky/automobile-dataset}
Blood Pressure Data

This data set contains blood pressure readings for patients before and after a certain treatment or intervention.

researchpy repo
Crop Yield Data

This data set contains information about crop yields.

researchpy repo
}
\usage{
df_crop_yield
}
\description{
This data set contains various automotive information including engine location, dimensions, weight, engine type, number of cylinders, and other specifications.
}
\keyword{datasets}
