% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FUNCTIONS_TRAIN_TEST.R
\name{result_confusion_performance}
\alias{result_confusion_performance}
\title{Plot performance of confusion matrix for different cut off points}
\usage{
result_confusion_performance(
  observed,
  predicted,
  step = 0.1,
  base_size = 10,
  title = ""
)
}
\arguments{
\item{observed}{vector of observed outcomes}

\item{predicted}{vector of predicted outcome probability}

\item{step}{stepping for tested cut values}

\item{base_size}{base font size}

\item{title}{plot title}
}
\description{
Plot performance of confusion matrix for different cut off points
}
\examples{
df<-data.frame(matrix(.999,ncol=2,nrow=2))
correlation_martix<-as.matrix(df)
diag(correlation_martix)<-1
df<-generate_correlation_matrix(correlation_martix,nrows=1000)
df$X1<-ifelse(abs(df$X1)<1,0,1)
df$X2<-abs(df$X2)
df$X2<-(df$X2-min(df$X2))/(max(df$X2)-min(df$X2))
result_confusion_performance(observed=round(abs(df$X1),0),
                             predicted=abs(df$X2),
                             step=.01)
result_confusion_performance(observed=c(1,2,3,1,2,3),predicted=abs(rnorm(6,0,sd=.1)))
}
\keyword{functions}
